### 2021-03-14 #3109 @arfon
@PacoCosta - this PR should fix your paper compilation https://github.com/SWE-UniStuttgart/Qlunc/pull/105
