### 2021-03-15 #3109 @PacoCosta
Dear Arfon,

Thanks for your comments.
Do you know how many time will take the complete process before publishing?
Thanks for your time and effort.

BR,
Paco
