### 2021-03-15 #3109 @arfon
> Do you know how many time will take the complete process before publishing?

I don't sorry. JOSS reviews usually take up to two months (but sometimes more), but this is once an editor has been assigned. I would hope you will have an editor assigned in the next few weeks.
