### 2021-03-25 09596f7
fixed TIA in plots
closes #84
