### 2021-03-25 057bb7a
STDV calculations in scanner
