### 2021-03-26 061edd2
Update Qlunc_Classes.py
delete stdv of cartesian coordinates in scanner class
