+++ Main/Qlunc_inputs.yml
@@ -20,9 +20,15 @@ Components:
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: VAD          # Type of the scanning head: VAD or SCAN (Scanning Lidar)
- 
+    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
+    
+    # STDV in lidar angles and focus distance
+    stdv focus distance: 0   # [m]
+    stdv Cone angle: 0       # [deg]
+    stdv Azimuth: 15          # [deg]   
+    
     # VAD Pattern:
+<<<<<<< Updated upstream
 
     Focus distance: [60]   # [m]
     Cone angle: [30]       # [deg]
@@ -32,15 +38,331 @@ Components:
     stdv Cone angle: 0     # [deg]
     stdv Azimuth: 15        # [deg]    
     
+=======
+    Focus distance: [40]   # [m]
+    Cone angle: [30]       # [deg]
+    Azimuth: [0,360,360]     # [deg] [origin,end,steps]
+    
+>>>>>>> Stashed changes
     # Scanning Lidar Pattern:
-    x: [40,40,40,40,40]        # [m]
-    y: [0,20,20,-20,-20]      # [m]
-    z: [0,20,-20,20,-20]      # [m]
-    
-    stdv x: 0.3               # [m]
-    stdv y: 1.6             # [m]
-    stdv z: 1.1             # [m]
-    
[…]
