+++ Main/Qlunc_Plotting.py
@@ -17,9 +17,9 @@ plot_param={'axes_label_fontsize' : 25,
             'title_fontsize'      : 29,
             'suptitle_fontsize'   : 23,
             'legend_fontsize'     : 15,
-            'xlim'                : [-10,60],
-            'ylim'                : [-40,40],
-            'zlim'                : [0,115],
+            'xlim'                : [-45,45],
+            'ylim'                : [-45,45],
+            'zlim'                : [-60,45],
             'markersize'          : 5,
             'markersize_lidar'    : 9,
             'marker'              : '.r',
