+++ Main/Qlunc_inputs.yml
@@ -20,7 +20,7 @@ Components:
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
+    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
@@ -33,9 +33,9 @@ Components:
     Azimuth: [0,360,360]     # [deg] [origin,end,steps]
 
     # Scanning Lidar Pattern:
-    x: [5]    
-    y: [-5]   
-    z: [10] #[10,   -10, -10 , 10 ]
+    x: [10,-10,-10,10]
+    y: [-5,5,-5,5]   
+    z: [20,20,20,20] #[10,   -10, -10 , 10 ]
         
     # Uncertainty function:
     Uncertainty function: uopc.UQ_Scanner
