+++ Main/Qlunc_inputs.yml
@@ -19,13 +19,14 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
     stdv Cone angle: 0       # [deg]
-    stdv Azimuth: 15          # [deg]   
+    stdv Azimuth: 0          # [deg]   
     
     # VAD Pattern:
     Focus distance: [60]   # [m]
