+++ Main/Qlunc_inputs.yml
@@ -19,24 +19,24 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
+    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
+    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
-    stdv Cone angle: 0       # [deg]
-    stdv Azimuth: 0          # [deg]   
+    stdv Cone angle: 0      # [deg]
+    stdv Azimuth: 5         # [deg]   
     
     # VAD Pattern:
-    Focus distance: [60]   # [m]
-    Cone angle: [30]       # [deg]
-    Azimuth: [0,360,360]     # [deg] [origin,end,steps]
+    Focus distance: [20]   # [m]
+    Cone angle: [5]       # [deg]
+    Azimuth: [0,360,15]     # [deg] [origin,end,steps]
 
     # Scanning Lidar Pattern:
     x: [10,-10,-10,10]
     y: [-5,5,-5,5]   
-    z: [20,20,20,20] #[10,   -10, -10 , 10 ]
+    z: [40,40,40,40] #[10,   -10, -10 , 10 ]
         
     # Uncertainty function:
     Uncertainty function: uopc.UQ_Scanner
@@ -130,4 +130,4 @@ Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
   Photodetector noise: False
-
+  Save Scanning Coordinates: True
