+++ UQ_Functions/UQ_Optics_Classes.py
@@ -10,17 +10,17 @@ Here we calculate the uncertainties related with components in the `optics`
 module. 
 
     
-   - noise definintions (reference in literature)
+   - noise definitions (reference in literature)
    
  
 """
 import os
-os.chdir('../Utils')
-import Scanning_patterns as SP
-from Qlunc_Help_standAlone import *
-os.chdir('../Main')
+
 import pdb
-#%% TELESCOPE:
+from Utils.Qlunc_ImportModules import *
+from Utils import Qlunc_Help_standAlone as SA
+from Utils import Scanning_patterns as SP
+
 # NOT IMPLEMENTED
 #==============================================================================
 # def UQ_Telescope(Lidar, Atmospheric_Scenario,cts):
@@ -54,14 +54,14 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
       [np.sin(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))*np.cos(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep)),  np.sin(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))*np.sin(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))*np.sin(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))+np.cos(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))*np.cos(np.deg2rad(Lidar.lidar_inputs.roll_error_dep)),  np.sin(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))*np.sin(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))*np.cos(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))-np.cos(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))*np.sin(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))],
       [       -np.sin(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))                                             ,                      np.cos(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))*np.sin(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))                                                                                                                                            ,                                                                np.cos(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))*np.cos(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))]]
     stdv_param1=Lidar.optics.scanner.stdv_focus_dist
-    stdv_param2=Lidar.optics.scanner.stdv_cone_angle
-    stdv_param3=Lidar.optics.scanner.stdv_azimuth
+    stdv_param2=np.deg2rad(Lidar.optics.scanner.stdv_cone_angle)
+    stdv_param3=np.deg2rad(Lidar.optics.scanner.stdv_azimuth)
     
     # Differentiate between 'VAD' or 'Scanning' lidar depending on user's choice:
     if Qlunc_yaml_inputs['Components']['Scanner']['Type']=='VAD':
         param1=Lidar.optics.scanner.focus_dist
-        param2=Lidar.optics.scanner.cone_angle
-        param3=Lidar.optics.scanner.azimuth      
+        param2=np.deg2rad(Lidar.optics.scanner.cone_angle)
+        param3=np.deg2rad(Lidar.optics.scanner.azimuth)
 
     elif Qlunc_yaml_inputs['Components']['Scanner']['Type']=='SCAN':
[…]
