### 2021-03-29 4c9034b
Update UQ_Power_Classes.py
