### 2021-03-29 8aa39b0
Update Scanning_patterns.py
