### 2021-03-29 b7d5d0d
Solving directories and plotting script
