+++ UQ_Functions/UQ_Optics_Classes.py
@@ -17,10 +17,15 @@ module.
 import os
 
 import pdb
+# print('Optics: '+os.getcwd())
+# pdb.set_trace()
 from Utils.Qlunc_ImportModules import *
 from Utils import Qlunc_Help_standAlone as SA
 from Utils import Scanning_patterns as SP
-
+from Utils import Qlunc_Plotting as QPlot
+print('SecionOptics: '+os.getcwd())
+# pdb.set_trace()
+# from  Main.Qlunc_Instantiate import *
 # NOT IMPLEMENTED
 #==============================================================================
 # def UQ_Telescope(Lidar, Atmospheric_Scenario,cts):
@@ -172,6 +177,10 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     #     file.close()   
         
     Final_Output_UQ_Scanner={'Simu_Mean_Distance':SimMean_DISTANCE,'STDV_Distance':StdvMean_DISTANCE,'MeasPoint_Coordinates':Coord,'NoisyMeasPoint_Coordinates':Noisy_Coord}
+    #%% Plotting
+    pdb.set_trace()
+    # if Qlunc_yaml_inputs['Flags']['Scanning Pattern']:
+    QPlot.plotting(Lidar,Qlunc_yaml_inputs,Final_Output_UQ_Scanner,True,False)
     return Final_Output_UQ_Scanner
 
 #%% Optical circulator:
@@ -199,7 +208,7 @@ def sum_unc_optics(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     except:
         Optical_circulator_Uncertainty = None
         print('No optical circulator in calculations!')
-           
     Uncertainty_Optics_Module=SA.unc_comb(List_Unc_optics)
-    Final_Output_UQ_Optics = {'Uncertainty_Optics':Uncertainty_Optics_Module,'Uncertainty_PointingAccuracy':[Scanner_Uncertainty[0],Scanner_Uncertainty[1]]}
+    Final_Output_UQ_Optics = {'Uncertainty_Optics':Uncertainty_Optics_Module,'Mean_error_PointingAccuracy':Scanner_Uncertainty['Simu_Mean_Distance'],'Stdv_PointingAccuracy':Scanner_Uncertainty['STDV_Distance']}
     return Final_Output_UQ_Optics
+
