+++ Main/Qlunc_inputs.yml
@@ -19,9 +19,9 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
+    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
@@ -130,4 +130,4 @@ Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
   Photodetector noise: False
-  Save Scanning Coordinates: True
+  Save Scanning Coordinates: False

[…]
