### 2021-03-29 0485927
Plotting function and finishes arranging directories
closes #63
