+++ UQ_Functions/UQ_Photonics_Classes.py
@@ -18,9 +18,10 @@ module.
 
 from Utils.Qlunc_ImportModules import *
 from Utils import Qlunc_Help_standAlone as SA
+from Utils import Qlunc_Plotting as QPlot
 
 #%% PHOTODETECTOR:
-def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts):
+def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     UQ_Photodetector.Thermal_noise      = []
     UQ_Photodetector.SNR_thermal_noise  = []
     UQ_Photodetector.Shot_noise         = []
@@ -64,6 +65,9 @@ def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts):
     
     UQ_Photodetector.UQ_Photo=list(SA.flatten(UQ_Photodetector.UQ_Photo))
     Final_Output_UQ_Photo={'Uncertainty_Photodetector':UQ_Photodetector.UQ_Photo,'SNR_data_photodetector':SNR_data}      
+    
+    # Plotting:
+    QPlot.plotting(Lidar,Qlunc_yaml_inputs,Final_Output_UQ_Photo,False,Qlunc_yaml_inputs['Flags']['Photodetector noise'])
     return Final_Output_UQ_Photo
 
 
