+++ TestFiles_Qlunc/WorkingExample_Qlunc_Instantiate.py
@@ -20,27 +20,29 @@ could be done by instantiating their classes:
         6) Lidar device instance is created and modules, containing the 
            different components are included in the lidar architecture 
       
- """   
+ """ 
+import os
+import pdb
+
+os.chdir('../')
+pdb.set_trace()
+# importing  uncertainty functions
+import UQ_Functions.UQ_Photonics_Classes as uphc,UQ_Functions.UQ_Optics_Classes as uopc, UQ_Functions.UQ_Power_Classes as upwc,UQ_Functions.UQ_Lidar_Classes as ulc
+from Utils.Qlunc_ImportModules import *
+  
 #%% Running Qlunc_Classes.py:
-from WorkingExample_Qlunc_Classes import *
 
 # Reading data from the yaml file:
-with open (r'./WorkingExample_yaml_inputs_file.yml') as file: # WHere the yaml file is in order to get the input data
+with open (r'./TestFiles_Qlunc/WorkingExample_yaml_inputs_file.yml') as file: # WHere the yaml file is in order to get the input data
     Qlunc_yaml_inputs={}
     docs = yaml.load_all(file, Loader=yaml.FullLoader)
     for doc in docs:      
         for k, v in doc.items():           
             Qlunc_yaml_inputs.setdefault(k,v)  # save a dictionary with the data coming from yaml file 
 
-exec(open(Qlunc_yaml_inputs['Main_directory']+'/WorkingExample_Qlunc_Classes.py').read())   # Execute Qlunc_Classes.py (creating classes for lidar 'objects')
+# Execute Qlunc_Classes.py (creating classes for lidar 'objects')
+exec(open(Qlunc_yaml_inputs['Main_directory']+'/TestFiles_Qlunc/WorkingExample_Qlunc_Classes.py').read())   # 
 #%%%%%%%%%%%%%%%%% INPUTS %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-## FLAG inputs: ###############################################################
-flags.flag_plot_pointing_accuracy_unc    = Qlunc_yaml_inputs['Flags']['Pointing accuracy uncertainty']   # Pointing accuracy uncertainty - Keep False
-flags.flag_plot_measuring_points_pattern = Qlunc_yaml_inputs['Flags']['Scanning Pattern']  # Pattern of measuring points
-flags.flag_plot_photodetector_noise      = Qlunc_yaml_inputs['Flags']['Photodetector noise']  # Photodetector noise: shot noise, dark current noise, thermal noise as a function of the photodetector input signal power.
-
-
 
 ## Optics components and Module: ##############################################
 
@@ -52,39 +54,41 @@ flags.flag_plot_photodetector_noise      = Qlunc_yaml_inputs['Flags']['Photodete
 Scanner           = scanner(name            = Qlunc_yaml_inputs['Components']['Scanner']['Name'],           # Introduce your scanner name.
                             scanner_type    = Qlunc_yaml_inputs['Components']['Scanner']['Type'],
                             origin          = Qlunc_yaml_inputs['Components']['Scanner']['Origin'],         # Origin (coordinates of the lidar deployment).
+                            pattern         = Qlunc_yaml_inputs['Components']['Scanner']['Pattern'],
                             sample_rate     = Qlunc_yaml_inputs['Components']['Scanner']['Sample rate'],    # for now introduce it in [degrees].
                            
                            # This values for focus distance, cone_angle and azimuth define a typical VAD scanning sequence:
[…]
