+++ TestFiles_Qlunc/WorkingExample_Qlunc_Plotting.py
@@ -7,69 +7,66 @@ Francisco Costa García
 University of Stuttgart(c) 
 
 """
-from WorkingExample_Qlunc_Instantiate import *
+#%% import packages:
+from Utils.Qlunc_ImportModules import *
 
 #%% Plotting:
-
-# Plotting general parameters:
-plot_param={'axes_label_fontsize' : 25,
-            'textbox_fontsize'    : 14,
-            'title_fontsize'      : 29,
-            'suptitle_fontsize'   : 23,
-            'legend_fontsize'     : 15,
-            'xlim'                : [-50,50],
-            'ylim'                : [-50,50],
-            'zlim'                : [0,150],
-            'markersize'          : 5,
-            'markersize_lidar'    : 9,
-            'marker'              : '.',
-            'markerTheo'          : '.r',
-            'tick_labelrotation'  : 45,
-            'Qlunc_version'       : 'Qlunc Version - 0.91'
-            }
-
-##############    Plotting scanner measuring points pattern and their uncertainty #######################
-if flags.flag_plot_measuring_points_pattern:
-    Scanner_Data1 = Lidar.optics.scanner.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs) # Calling Scanner uncertainty to plot the graphics
-
+def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,flag_plot_photodetector_noise):
+    # Ploting general parameters:
+    plot_param={'axes_label_fontsize' : 25,
+                'textbox_fontsize'    : 14,
+                'title_fontsize'      : 29,
+                'suptitle_fontsize'   : 23,
+                'legend_fontsize'     : 15,
+                'xlim'                : [-45,45],
+                'ylim'                : [-45,45],
+                'zlim'                : [-40,45],
+                'markersize'          : 5,
+                'markersize_lidar'    : 9,
+                'marker'              : '.r',
+                'markerTheo'          : '.b',
+                'tick_labelrotation'  : 45,
+                'Qlunc_version'       : 'Qlunc Version - 0.9'
+                }
+        
+    ##############    Ploting scanner measuring points pattern #######################
+    if flag_plot_measuring_points_pattern:              
+        # Creating the figure and the axes
+        fig,axs4 = plt.subplots()  
+        axs4=plt.axes(projection='3d')
+        # pdb.set_trace()
+        # Plotting
+        axs4.plot([Lidar.optics.scanner.origin[0]],[Lidar.optics.scanner.origin[1]],[Lidar.optics.scanner.origin[2]],'ob',label='{} coordinates [{},{},{}]'.format(Lidar.LidarID,Lidar.optics.scanner.origin[0],Lidar.optics.scanner.origin[1],Lidar.optics.scanner.origin[2]),markersize=plot_param['markersize_lidar'])
+        axs4.plot(Data['MeasPoint_Coordinates'][0],Data['MeasPoint_Coordinates'][1],Data['MeasPoint_Coordinates'][2],plot_param['markerTheo'],markersize=plot_param['markersize'],label='Theoretical measuring point')
[…]

[…]
