+++ Main/Qlunc_inputs.yml
@@ -21,27 +21,23 @@ Components:
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
     Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
     stdv Cone angle: 10      # [deg]
     stdv Azimuth: 0         # [deg]   
-
-    
+   
     # VAD Pattern:
     Focus distance: [58.30951895]   # [m]
     Cone angle: [46.68614334]       # [deg]
     Azimuth: [0,360,90]     # [deg] [origin,end,steps]
 
     # Scanning Lidar Pattern:
-
     x: [-40]
     y: [0]   
     z: [0] #[10,   -10, -10 , 10 ]
         
-
     # Uncertainty function:
     Uncertainty function: uopc.UQ_Scanner
     
@@ -119,7 +115,6 @@ Lidar:
 # Atmospheric_inputs to create the atmospheric scenarios
 
 Atmospheric_inputs:
-
   Name: 'Atmosphere_Yaml'
   TimeSeries    : False # This defines whether we are using a time series (True) or single values (False) to describe the atmosphere characteristics (T, H, rain and fog)   
   Atmos_TS_FILE : 'AtmosphericScenarios.csv'  # (... .csv)
