+++ Tutorials/yaml_inputs_file_1.yml
@@ -9,7 +9,7 @@
 # DIRECTORIES:------------------------------------------------------
 
 Main_directory: # Directory where Qlunc_Classes.py is
-  '../TestFiles_Qlunc'
+  './TestFiles_Qlunc'
 
 # Lidar values. Modules and components parameters: ------------------------------------------------------
 
