+++ Tutorials/yaml_inputs_file_2_1.yml
@@ -17,7 +17,7 @@ Components:
 #Optics: 
   Scanner:
     
-    Name: Scanner_Yaml2 # Provide an ID to your lidar:    
+    Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]
     Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0.
@@ -29,7 +29,7 @@ Components:
     stdv Azimuth: 0         # [deg] 
     
     # VAD Pattern:
-    Focus distance: [30]   # [m]
+    Focus distance: [40]   # [m]
     Cone angle: [20]       # [deg]
     Azimuth: [0,360,8]    # [deg]
         
@@ -42,14 +42,14 @@ Components:
     Uncertainty function: uopc.UQ_Scanner
     
   Optical Circulator:
-    Name: OC_Yaml2
+    Name: OC_Yaml
     Insertion loss: 2.1 # [dB]
     Uncertainty function: uopc.UQ_OpticalCirculator
   
 # Photonics  
   Optical Amplifier: 
   
-    Name: OA_Yaml2
+    Name: OA_Yaml
     
     # Noise figure in dB. Could be a well known value or a table.
     # If table (FigureNoise Vs Wavelength) is provided Qlunc 
@@ -67,7 +67,7 @@ Components:
 
   
   Photodetector:
-    Name: Photodetector_Yaml2
+    Name: Photodetector_Yaml
     Photodetector BandWidth: 380.e+6    # [Hz]            
     Load resistor: 50                   # [Ohm]
     Photodetector efficiency: .85       # Photodetector efficiency [-]
@@ -80,7 +80,7 @@ Components:
  
 # Lidar general inputs 
   Lidar general inputs:
-    Name: Lidar_gral_inputs_Yaml2
+    Name: Lidar_gral_inputs_Yaml
     Wavelength: 1550.e-9    # [m]
     Yaw error: 0            # [deg]
     Pitch error: 0          # [deg]
@@ -89,7 +89,7 @@ Components:
 # Modules:
 Modules:
   Optics Module:
-    Name: Optics_module_Yaml2
+    Name: Optics_module_Yaml
     Scanner: Scanner
     Optical circulator: Optical_circulator # Have to be the same name as the instance
     Laser: None
@@ -102,7 +102,7 @@ Modules:
 ---  
 # Lidar:
 Lidar:
-  Name: Caixa2
+  Name: Caixa1
   Photonics module: Photonics_Module       # Have to be the same name as the instance
   Optics module: Optics_Module             # Have to be the same name as the instance
   Power module: None                       # Have to be the same name as the instance
@@ -115,7 +115,7 @@ Lidar:
 # Atmospheric_inputs to create the atmospheric scenarios
 
 Atmospheric_inputs:
-  Name: 'Atmosphere_Yaml2'
+  Name: 'Atmosphere_Yaml'
   TimeSeries    : False # This defines whether we are using a time series (True) or single values (False) to describe the atmosphere characteristics (T, H, rain and fog)   
   Atmos_TS_FILE : 'AtmosphericScenarios.csv'  
   Temperature: [300] # [K] Have to put the same number of elements for temperature and humidity. Always in paired values   
@@ -127,6 +127,6 @@ Atmospheric_inputs:
 # Plotting:------------------------------------------------------------
 Flags:
   Pointing accuracy uncertainty: False
-  Scanning Pattern: True
-  Photodetector noise: True
+  Scanning Pattern: False
+  Photodetector noise: False
   Save Scanning Coordinates: False
