+++ Tutorials/yaml_inputs_file_2_2.yml
@@ -106,7 +106,7 @@ Lidar:
   Photonics module: Photonics_Module       # Have to be the same name as the instance
   Optics module: Optics_Module2             # Have to be the same name as the instance
   Power module: None                       # Have to be the same name as the instance
-  Lidar inputs: Lidar_inputs2               # Have to be the same name as the instance
+  Lidar inputs: Lidar_inputs               # Have to be the same name as the instance
   Uncertainty function: ulc.sum_unc_lidar
 
 ---
