### 2021-03-30 de68240
Create yaml_inputs_file_2_1.yml
