+++ UQ_Functions/UQ_Photonics_Classes.py
@@ -19,7 +19,8 @@ module.
 from Utils.Qlunc_ImportModules import *
 from Utils import Qlunc_Help_standAlone as SA
 from Utils import Qlunc_Plotting as QPlot
-
+import os
+import pdb
 #%% PHOTODETECTOR:
 def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     UQ_Photodetector.Thermal_noise      = []
@@ -73,14 +74,15 @@ def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
 
 
 #%% OPTICAL AMPLIFIER:
-def UQ_Optical_amplifier(Lidar,Atmospheric_Scenario,cts): 
-
+def UQ_Optical_amplifier(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs): 
+    print(os.getcwd())
+    pdb.set_trace()
 # Obtain SNR from figure noise or pass directly numerical value:
     if isinstance (Lidar.photonics.optical_amplifier.NoiseFig, numbers.Number): #If user introduces a number or a table of values
         FigureNoise=[(Lidar.photonics.optical_amplifier.NoiseFig)]*len(Atmospheric_Scenario.temperature) #Figure noise vector        
         
         # ASE noise:
-        UQ_Optical_amplifier = [np.array([10*np.log10((10**(FigureNoise[0]/10))*cts.h*(cts.c/Lidar.lidar_inputs.Wavelength)*10**(Lidar.photonics.optical_amplifier.Gain/10))]*len(Atmospheric_Scenario.temperature))] 
+        UQ_Optical_amplifier = [np.array([10*np.log10((10**(FigureNoise[0]/10))*cts.h*(cts.c/Lidar.lidar_inputs.Wavelength)*10**(Lidar.photonics.optical_amplifier.OA_Gain/10))]*len(Atmospheric_Scenario.temperature))] 
     else:
         NoiseFigure_DATA = pd.read_csv(Lidar.photonics.optical_amplifier.NoiseFig,delimiter=';',decimal=',') #read from a .csv file variation of dB with wavelength (for now just with wavelength)    
         figure_noise_INT  = itp.interp1d(NoiseFigure_DATA.iloc[:,0],NoiseFigure_DATA.iloc[:,1],kind='cubic',fill_value="extrapolate")# First column wavelength,second column SNR in dB
@@ -88,7 +90,7 @@ def UQ_Optical_amplifier(Lidar,Atmospheric_Scenario,cts):
         FigureNoise=(NoiseFigure_VALUE.tolist())
         
         # ASE noise:
-        UQ_Optical_amplifier = [np.array([10*np.log10((10**(FigureNoise/10))*cts.h*(cts.c/Lidar.lidar_inputs.Wavelength)*10**(Lidar.photonics.optical_amplifier.Gain/10))]*len(Atmospheric_Scenario.temperature)) ]
+        UQ_Optical_amplifier = [np.array([10*np.log10((10**(FigureNoise/10))*cts.h*(cts.c/Lidar.lidar_inputs.Wavelength)*10**(Lidar.photonics.optical_amplifier.OA_Gain/10))]*len(Atmospheric_Scenario.temperature)) ]
 
     Final_Output_UQ_Optical_Amplifier={'Uncertainty_OpticalAmp':UQ_Optical_amplifier}
     return Final_Output_UQ_Optical_Amplifier
@@ -103,7 +105,7 @@ def sum_unc_photonics(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
         Photodetector_Uncertainty=None
[…]
