+++ Main/Qlunc_inputs.yml
@@ -19,7 +19,7 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
@@ -57,7 +57,7 @@ Components:
     # with the lidar nominal wavelength. Interpolation param. could 
     # be tunned in Qlunc_Photonics_Classes --> UQ_Optical_amplifier
     
-    Optical amplifier noise figure: ../metadata/NoiseFigure.csv  # Either a well-known value or the directory where the file is.
+    Optical amplifier noise figure: ./metadata/NoiseFigure.csv  # Relative Directory to where Noise figure data is. Either a well-known value or the directory where the file is.
     
     # Optical amplifier gain in dB
     Optical amplifier gain: 30
