### 2021-03-30 9e9a1b1
Include Total SNR in photodetector plot
closes #120
