+++ Main/Qlunc_inputs.yml
@@ -71,11 +71,11 @@ Components:
     Photodetector BandWidth: 380.e+6    # [Hz]            
     Load resistor: 50                   # [Ohm]
     Photodetector efficiency: .85       # Photodetector efficiency [-]
-    Dark current: 5.e-9                 # In [A]. Dark current in the photodetector.
+    Dark current: 5.e-1                 # In [A]. Dark current in the photodetector.
     Photodetector signalP: 1.e-3        # [W]
     Power interval: [0,1000,.001]       # [W]
     Gain TIA: 5.e+3                     # [dB] Value or 'None' if no TIA in the photodetector
-    V Noise TIA: 160.e-6 #None #               # [V] Value or 'None' if no TIA in the photodetector
+    V Noise TIA:  160.e-6 #None #             # [V] Value or 'None' if no TIA in the photodetector
     Uncertainty function: uphc.UQ_Photodetector # Python module! (You develop a function and save it as a python module)
  
 # Lidar general inputs 
