+++ Utils/Qlunc_Plotting.py
@@ -44,13 +44,11 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
                 'Qlunc_version'       : 'Qlunc Version - 0.9'
                 }
         
-    ##############    Ploting scanner measuring points pattern #######################
+##################    Ploting scanner measuring points pattern #######################
     if flag_plot_measuring_points_pattern:              
-        # Creating the figure and the axes
+        # Plotting
         fig,axs4 = plt.subplots()  
         axs4=plt.axes(projection='3d')
-        # pdb.set_trace()
-        # Plotting
         axs4.plot([Lidar.optics.scanner.origin[0]],[Lidar.optics.scanner.origin[1]],[Lidar.optics.scanner.origin[2]],'ob',label='{} coordinates [{},{},{}]'.format(Lidar.LidarID,Lidar.optics.scanner.origin[0],Lidar.optics.scanner.origin[1],Lidar.optics.scanner.origin[2]),markersize=plot_param['markersize_lidar'])
         axs4.plot(Data['MeasPoint_Coordinates'][0],Data['MeasPoint_Coordinates'][1],Data['MeasPoint_Coordinates'][2],plot_param['markerTheo'],markersize=plot_param['markersize'],label='Theoretical measuring point')
         axs4.plot(Data['NoisyMeasPoint_Coordinates'][0],Data['NoisyMeasPoint_Coordinates'][1],Data['NoisyMeasPoint_Coordinates'][2],plot_param['marker'],markersize=plot_param['markersize'],label='Distance error [m] = {0:.3g}$\pm${1:.3g}'.format(np.mean(Data['Simu_Mean_Distance']),np.mean(Data['STDV_Distance'])))
@@ -58,16 +56,14 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
         # Setting labels, legend, title and axes limits:
         axs4.set_xlabel('x [m]',fontsize=plot_param['axes_label_fontsize'])#,orientation=plot_param['tick_labelrotation'])
         axs4.set_ylabel('y [m]',fontsize=plot_param['axes_label_fontsize'])#,orientation=plot_param['tick_labelrotation'])
-        axs4.set_zlabel('z [m]',fontsize=plot_param['axes_label_fontsize'])
-    
-        
+        axs4.set_zlabel('z [m]',fontsize=plot_param['axes_label_fontsize'])        
         axs4.set_title('Lidar pointing accuracy ['+Qlunc_yaml_inputs['Components']['Scanner']['Type']+']',fontsize=plot_param['title_fontsize'])
         axs4.legend()
         axs4.set_xlim3d(plot_param['xlim'][0],plot_param['xlim'][1])
         axs4.set_ylim3d(plot_param['ylim'][0],plot_param['ylim'][1])
         axs4.set_zlim3d(plot_param['zlim'][0],plot_param['zlim'][1])
        
-    ###########   Plot photodetector noise   #############################       
+###############   Plot photodetector noise   #############################       
     if flag_plot_photodetector_noise:
         # Quantifying uncertainty from photodetector and interval domain for the plot Psax is define in the photodetector class properties)
         Psax=10*np.log10(Lidar.photonics.photodetector.Power_interval) 
@@ -76,8 +72,7 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
         fig,ax=plt.subplots()
         label0=['Shot SNR','Thermal SNR','Dark current SNR','TIA SNR']
         i_label=0
-        for i in Data['SNR_data_photodetector']:
-            
[…]
