### 2021-03-30 fc1795e
Update Scanning_patterns.py
