+++ Utils/Scanning_patterns.py
@@ -9,7 +9,7 @@ from Utils.Qlunc_ImportModules import *
 
 #%% Lissajous pattern
 def lissajous_pattern(size_x,size_y,size_z,q,p):
-        """
+    """
     Plotting. Location: .Utils/Qlunc_plotting.py
     
     Parameters
