+++ Main/Qlunc_inputs.yml
@@ -8,8 +8,8 @@
 
 # DIRECTORIES:------------------------------------------------------
 
-Main_directory: # Directory where Qlunc_Classes.py is
-  '.'
+Main_directory: '.' # Directory where Qlunc_Classes.py is
+  
 
 # Lidar values. Modules and components parameters: ------------------------------------------------------
 
@@ -19,7 +19,7 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
