+++ UQ_Functions/UQ_Optics_Classes.py
@@ -169,6 +169,7 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     # file.close()   
         
     Final_Output_UQ_Scanner={'Simu_Mean_Distance':SimMean_DISTANCE,'STDV_Distance':StdvMean_DISTANCE,'MeasPoint_Coordinates':Coord,'NoisyMeasPoint_Coordinates':Noisy_Coord}
+    Lidar.lidar_inputs.dataframe['Scanner']=Final_Output_UQ_Scanner
     # Plotting
     QPlot.plotting(Lidar,Qlunc_yaml_inputs,Final_Output_UQ_Scanner,Qlunc_yaml_inputs['Flags']['Scanning Pattern'],False)
     return Final_Output_UQ_Scanner
@@ -177,6 +178,8 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
 def UQ_OpticalCirculator(Lidar,Atmospheric_Scenario,cts):
     Optical_Circulator_Uncertainty = [Lidar.optics.optical_circulator.insertion_loss]
     Final_Output_UQ_Optical_Circulator={'Optical_Circulator_Uncertainty':Optical_Circulator_Uncertainty}
+    Lidar.lidar_inputs.dataframe['Optical circulator']=Final_Output_UQ_Optical_Circulator
+
     return Final_Output_UQ_Optical_Circulator
 
 #%% Sum of uncertainties in `optics` module: 
@@ -200,5 +203,6 @@ def sum_unc_optics(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
         print('No optical circulator in calculations!')
     Uncertainty_Optics_Module=SA.unc_comb(List_Unc_optics)
     Final_Output_UQ_Optics = {'Uncertainty_Optics':Uncertainty_Optics_Module,'Mean_error_PointingAccuracy':Scanner_Uncertainty['Simu_Mean_Distance'],'Stdv_PointingAccuracy':Scanner_Uncertainty['STDV_Distance']}
+    Lidar.lidar_inputs.dataframe['Optics Module']=Final_Output_UQ_Optics
     return Final_Output_UQ_Optics
 
