+++ UQ_Functions/UQ_Lidar_Classes.py
@@ -40,5 +40,7 @@ def sum_unc_lidar(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     print('Processing lidar uncertainties...')
     Uncertainty_Lidar=SA.unc_comb(List_Unc_lidar)
     Final_Output_Lidar_Uncertainty = {'Lidar_Uncertainty':Uncertainty_Lidar}    
+    Lidar.lidar_inputs.dataframe['Lidar']=Final_Output_Lidar_Uncertainty
+
     print('Lidar uncertainty done')
     return Final_Output_Lidar_Uncertainty
