### 2021-04-01 7942e78
Getting dictionary final results
