+++ UQ_Functions/UQ_Lidar_Classes.py
@@ -19,20 +19,20 @@ def sum_unc_lidar(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     List_Unc_lidar = []
     try: # ecah try/except evaluates wether the component is included in the module, therefore in the calculations
 #        if Photodetector_Uncertainty not in locals():
-        Photonics_Uncertainty=Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+        Photonics_Uncertainty,DataFrame=Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
         List_Unc_lidar.append(Photonics_Uncertainty['Uncertainty_Photonics'])
     except:
         Photonics_Uncertainty=None
         print('No photonics module in calculations!')
     try:
-        Optics_Uncertainty=Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+        Optics_Uncertainty,DataFrame=Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
         Optics_Uncertainty=np.ndarray.tolist(Optics_Uncertainty['Uncertainty_Optics'])*len(Atmospheric_Scenario.temperature)
         List_Unc_lidar.append(np.array([Optics_Uncertainty]))
     except:
         Optics_Uncertainty=None
         print('No optics module in calculations!')
     try:
-        Power_Uncertainty=Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
+        Power_Uncertainty,DataFrame=Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
         List_Unc_lidar.append(Power_Uncertainty['Uncertainty_Power']*len(Atmospheric_Scenario.temperature))
     except:
         Power_Uncertainty=None
