### 2021-04-01 0f37a1d
Update UQ_Lidar_Classes.py
closes #71
