+++ TestFiles_Qlunc/WorkingExample_Qlunc_Instantiate.py
@@ -142,7 +142,7 @@ Lidar = lidar(name         = Qlunc_yaml_inputs['Lidar']['Name'],                
 Atmospheric_TimeSeries = Qlunc_yaml_inputs['Atmospheric_inputs']['TimeSeries'] # This defines whether we are using a time series (True) or single values (False) to describe the atmosphere (T, H, rain and fog) 
                                                                            # If so we obtain a time series describing the noise implemented in the measurement.
 if Atmospheric_TimeSeries:
-    Atmos_TS_FILE           = '../metadata/AtmosphericData/'+Qlunc_yaml_inputs['Atmospheric_inputs']['Atmos_TS_FILE']
+    Atmos_TS_FILE           = './metadata/AtmosphericData/'+Qlunc_yaml_inputs['Atmospheric_inputs']['Atmos_TS_FILE']
     AtmosphericScenarios_TS = pd.read_csv(Atmos_TS_FILE,delimiter=';',decimal=',')
     Atmospheric_inputs = {
                           'temperature' : list(AtmosphericScenarios_TS.loc[:,'T']),    # [K]
