+++ Tutorials/yaml_inputs_file_1 - Copy.yml
@@ -1,132 +1,0 @@
-# Francisco Costa García
-# University of Stuttgart(c) 
-
-# Inputs file for Qlunc (Quantification lidar uncertainties). 
-# Do not change definition names or vary the structure of the file!! Just change parameter values. 
-# Users can find a yaml format template in the repository '/Main/Template_yaml_inputs_file.yml'
-
-
-# DIRECTORIES:------------------------------------------------------
-
-Main_directory: # Directory where Qlunc_Classes.py is
-  './TestFiles_Qlunc'
-
-# Lidar values. Modules and components parameters: ------------------------------------------------------
-
-Components:
-#Optics: 
-  Scanner:
-    
-    Name: Scanner_Yaml # Provide an ID to your lidar:    
-    Origin: [0,0,0]  # Lidar position in meters: [x,y,z]
-    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case no predefined pattern is used
-    Sample rate: 0   # Lidar sample rate. In meters. Use 0.
-    Type: VAD          # Type of the scanning head: VAD or SCAN (Scanning Lidar)
- 
-    # STDV in lidar angles and focus distance
-    stdv focus distance: 0   # [m]
-    stdv Cone angle: 10      # [deg]
-    stdv Azimuth: 0         # [deg] 
-    
-    # VAD Pattern:
-    Focus distance: [40]   # [m]
-    Cone angle: [20]       # [deg]
-    Azimuth: [0,360,8]    # [deg]
-        
-    # Scanning Lidar Pattern:
-    x: [20,20,20]        # [m]
-    y: [0,0,0]      # [m]
-    z: [20,0,-20]      # [m]
-
-    # Uncertainty function:
-    Uncertainty function: uopc.UQ_Scanner
-    
-  Optical Circulator:
-    Name: OC_Yaml
-    Insertion loss: 2.1 # [dB]
-    Uncertainty function: uopc.UQ_OpticalCirculator
-  
-# Photonics  
-  Optical Amplifier: 
-  
-    Name: OA_Yaml
-    
-    # Noise figure in dB. Could be a well known value or a table.
-    # If table (FigureNoise Vs Wavelength) is provided Qlunc 
-    # interpolates to obtain a most accurate value according 
-    # with the lidar nominal wavelength. Interpolation param. could 
-    # be tunned in Qlunc_Photonics_Classes --> UQ_Optical_amplifier
-    
-    Optical amplifier noise figure: ../metadata/NoiseFigure.csv 
-    
-    # Optical amplifier gain in dB
-    Optical amplifier gain: 30
-    
-    # Uncertainty function quantifying the optical amplifier uncertainty
-    Uncertainty function: uphc.UQ_Optical_amplifier
-
-  
-  Photodetector:
-    Name: Photodetector_Yaml
-    Photodetector BandWidth: 380.e+6    # [Hz]            
-    Load resistor: 50                   # [Ohm]
-    Photodetector efficiency: .85       # Photodetector efficiency [-]
-    Dark current: 5.e-9                 # In [A]. Dark current in the photodetector.
-    Photodetector signalP: 1.e-3        # [W]
-    Power interval: [0,1000,.001]       # [W]
-    Gain TIA: 5.e+3                     # [dB]
-    V Noise TIA: 160.e-6                # [V]
-    Uncertainty function: uphc.UQ_Photodetector
- 
-# Lidar general inputs 
-  Lidar general inputs:
-    Name: Lidar_gral_inputs_Yaml
-    Wavelength: 1550.e-9    # [m]
-    Yaw error: 0            # [deg]
-    Pitch error: 0          # [deg]
[…]
