+++ Tutorials/WorkingExample_Tutorial.py
@@ -1,0 +1,239 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jun 19 19:57:05 2020
+@author: fcosta
+
+Francisco Costa García
+University of Stuttgart(c) 
+
+###############################################################################
+###################### __ Classes __ ##########################################
+  
+Creating components, modules and lidar classes. 
+The command class is used to create digital "objects", representing the lidar 
+components, and store necessary information to calculate uncertainties coming 
+from hardware and data processing methods. 
+
+`unc_func` contains the functions calculating the uncertainty for each
+component, module and/or lidar system.
+It is defined as a python module, so users can define their own uncertainty 
+functions and implement them easily just pointing towards a different python 
+module. 
+
+Flags, general constants and atmospheric scenarios used along the calculations 
+are also treated as classes. User should instantiate all, but the general 
+constants.
+
+
+How is the code working?
+
+If user wants to implement another component/module s/he should create a class. 
+
+For example, we might want build up a lidar and user wants to include a 'power'
+module which, in turn contains an UPS (uninterruptible power supply) with 
+certain parameter values, e.g. output power and output voltage. Let's call them
+Pout and Vout. 
+
+1) Create the `classes`:
+    
+So we create the class `UPS`:
+    
+def UPS():
+    def __init__(self,PowerOutput,VoltageOutput,Uncertainty)
+        
+        self.Pout     = PowerOutput
+        self.Vout     = VoltageOutput
+        self.unc_func = Uncertainty
+
+And also the class `power`, including the UPS component:
+
+def power():
+    def __init__UPS (self, U_PowerSupply,Uncertainty)
+        self.UPS     = U_PowerSupply
+        self.unc_fun = Uncertainty 
+
+Then, the lidar is created (in this simple example) by including the module in 
+a lidar class:
+
+class lidar():
+      def __init__ (self, Power_Module,Uncertainty)
+          self.PowerMod = Power_Module
+          self.unc_fun  = Uncertainty
+
+Setting an atmospheric scenario is also needed:
+class atmos_sc():
+    def __init__ (self, Temperature, Humidity):
+        self.T = Temperature
+        self.H = Humidity
+    
+2) Instantiate the classes
+
+Instantiating the component class:
+UPS_1 = UPS(Pout     = 500, % in watts
+            Vout     = 24,  % in volts
+            unc_func = function calculating uncertainties)
+
+Instantiating the module class to create the `Power` object:
+Power_1 = power(Power_Module = UPS_1,
+                unc_func      = function calculating uncertainties) 
+        
+
+Instantiating the lidar class:
+Lidar_1 = lidar(Power_Module = Power_1,
+                unc_func     = function calculating uncertainties)
+
+So we have created a lidar digital twin with its first module, the `power` 
+module, which in turn contains a component, the uninterruptible power supply.
+
+Qlunc uses GUM (Guide to the expression of Uncertainties in Measurement) 
+suggestions to calculate uncertainty expansion.  
+  
+"""
+
+#%% Constants:
+class cts():
+    k = 1.38064852e-23 # Boltzman constant:[m^2 kg s^-2 K^-1]
+    h = 6.6207004e-34 # Plank constant [m^2 kg s^-1]
+    e = 1.60217662e-19 # electron charge [C]
+    c = 2.99792e8 #speed of light [m s^-1]
+    
+#%% LIDAR COMPONENTS
+  
+#Component Classes:
+class photodetector():
+    def __init__(self,name,Photo_BandWidth,Load_Resistor,Photo_efficiency,Dark_Current,Photo_SignalP,Power_interval,Gain_TIA,V_Noise_TIA,unc_func):
[…]
