+++ Main/Qlunc_inputs.yml
@@ -19,8 +19,8 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
-    Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
+    Sample rate: .03   # Lidar sample rate. In meters. Use 0 for now.
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
