+++ Utils/Qlunc_Plotting.py
@@ -70,12 +70,12 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
     
         # Plotting:
         fig,ax=plt.subplots()
-        label0=['Shot SNR','Thermal SNR','Dark current SNR','TIA SNR']
+        label0=['Shot SNR','Thermal SNR','Dark current SNR','Total SNR1','TIA SNR']
         i_label=0
         for i in Data['SNR_data_photodetector']:            
             ax.plot(Psax,Data['SNR_data_photodetector'][i][0],label=label0[i_label])  
             i_label+=1
-        ax.plot(Psax,Data['Total_SNR_data'],label='Total SNR')
+        # ax.plot(Psax,Data['Total_SNR_data'],label='Total SNR')
         ax.set_xlabel('Input Signal optical power [dBm] or [mW]',fontsize=plot_param['axes_label_fontsize'])
         ax.set_ylabel('SNR [dB]',fontsize=plot_param['axes_label_fontsize'])
         ax.legend(fontsize=plot_param['legend_fontsize'])
