### 2021-04-12 ff209b4
Total SNR calculations
