+++ UQ_Functions/UQ_Photonics_Classes.py
@@ -46,50 +46,83 @@ def UQ_Photodetector(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     """    
     
     UQ_Photodetector.Thermal_noise      = []
-    UQ_Photodetector.SNR_thermal_noise  = []
+    UQ_Photodetector.SNR_thermal  = []
     UQ_Photodetector.Shot_noise         = []
-    UQ_Photodetector.SNR_Shot_noise     = []
+    UQ_Photodetector.SNR_shot     = []
     UQ_Photodetector.Dark_current_noise = []
     UQ_Photodetector.SNR_DarkCurrent    = []
     UQ_Photodetector.UQ_Photo           = []
     UQ_Photodetector.SNR_TIA            = []
     UQ_Photodetector.TIA_noise          = []
-    SNR_data={}
-    Total_SNR_data=[]
+    
     R = Lidar.photonics.photodetector.Efficiency*cts.e*Lidar.lidar_inputs.Wavelength/(cts.h*cts.c)  #[A/W]  Responsivity
     UQ_Photodetector.Responsivity = (R) # this notation allows me to get Responsivity from outside of the function 
-    UQ_Photodetector.SNR_thermal_noise = [10*np.log10(((R**2)/(4*cts.k*Atmospheric_Scenario.temperature[0]*Lidar.photonics.photodetector.BandWidth/Lidar.photonics.photodetector.Load_Resistor))*(Lidar.photonics.photodetector.Power_interval/1000)**2)]
-    UQ_Photodetector.SNR_Shot_noise    = [10*np.log10(((R**2)/(2*cts.e*R*Lidar.photonics.photodetector.BandWidth))*(Lidar.photonics.photodetector.Power_interval)/1000)]
-    UQ_Photodetector.SNR_DarkCurrent   = [10*np.log10(((R**2)/(2*cts.e*Lidar.photonics.photodetector.DarkCurrent*Lidar.photonics.photodetector.BandWidth))*((Lidar.photonics.photodetector.Power_interval/1000)**2))]
-    for i in range(len(Atmospheric_Scenario.temperature)):
+    
+    # SNR calculations:
+    UQ_Photodetector.SNR_thermal_w = [((R**2)/(4*cts.k*Atmospheric_Scenario.temperature[0]*Lidar.photonics.photodetector.BandWidth/Lidar.photonics.photodetector.Load_Resistor))*(Lidar.photonics.photodetector.Power_interval/1000)**2]
+    UQ_Photodetector.SNR_shot_w    = [((R**2)/(2*cts.e*R*Lidar.photonics.photodetector.BandWidth))*(Lidar.photonics.photodetector.Power_interval)/1000]
+    UQ_Photodetector.SNR_DarkCurrent_w   = [((R**2)/(2*cts.e*Lidar.photonics.photodetector.DarkCurrent*Lidar.photonics.photodetector.BandWidth))*(Lidar.photonics.photodetector.Power_interval/1000)**2]
+    
+    UQ_Photodetector.SNR_thermal     = [10*np.log10(UQ_Photodetector.SNR_thermal_w)][0]
+    UQ_Photodetector.SNR_shot        = [10*np.log10(UQ_Photodetector.SNR_shot_w )][0]
+    UQ_Photodetector.SNR_DarkCurrent = [10*np.log10(UQ_Photodetector.SNR_DarkCurrent_w)][0]
+    
     
     # Photodetector Thermal noise:
+    for i in range(len(Atmospheric_Scenario.temperature)):
[…]
