+++ Utils/Qlunc_Plotting.py
@@ -70,7 +70,7 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
     
         # Plotting:
         fig,ax=plt.subplots()
-        label0=['Shot SNR','Thermal SNR','Dark current SNR','Total SNR1','TIA SNR']
+        label0=['Shot SNR','Thermal SNR','Dark current SNR','Total SNR','TIA SNR']
         i_label=0
         for i in Data['SNR_data_photodetector']:            
             ax.plot(Psax,Data['SNR_data_photodetector'][i][0],label=label0[i_label])  
