+++ Main/Qlunc_inputs.yml
@@ -19,19 +19,19 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
+    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
+    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar)
     
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
-    stdv Cone angle: 5      # [deg]
+    stdv Cone angle: 2      # [deg]
     stdv Azimuth: 0         # [deg]   
    
     # VAD Pattern:
-    Focus distance: [58.30951895]   # [m]
-    Cone angle: [46.68614334]       # [deg]
-    Azimuth: [0,360,90]     # [deg] [origin,end,steps]
+    Focus distance: [70]   # [m]
+    Cone angle: [30]       # [deg]
+    Azimuth: [0,360,45]     # [deg] [origin,end,steps]
 
     # Scanning Lidar Pattern:
     x: [-40]
@@ -128,5 +128,5 @@ Atmospheric_inputs:
 Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
-  Photodetector noise: True
+  Photodetector noise: False
   Save Scanning Coordinates: False
