+++ Utils/Qlunc_Plotting.py
@@ -33,9 +33,9 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
                 'title_fontsize'      : 29,
                 'suptitle_fontsize'   : 23,
                 'legend_fontsize'     : 15,
-                'xlim'                : [-45,45],
-                'ylim'                : [-45,45],
-                'zlim'                : [-40,45],
+                'xlim'                : [-60,60],
+                'ylim'                : [-60,60],
+                'zlim'                : [0,80],
                 'markersize'          : 5,
                 'markersize_lidar'    : 9,
                 'marker'              : '.r',
