+++ Main/Qlunc_inputs.yml
@@ -75,7 +75,7 @@ Components:
     Photodetector signalP: 1.e-6        # [W]
     Power interval: [0,1000,.001]       # [mW]
     Gain TIA: 5.e+3                     # [dB] Value or 'None' if no TIA in the photodetector
-    V Noise TIA:  160.e-6 # None #             # [V] Value or 'None' if no TIA in the photodetector
+    V Noise TIA:  None # 160.e-6 #             # [V] Value or 'None' if no TIA in the photodetector
     Uncertainty function: uphc.UQ_Photodetector # Python module! (You develop a function and save it as a python module)
  
 # Lidar general inputs 
