### 2021-04-16 cad360a
photodetector SNR total working
