+++ Main/Qlunc_inputs.yml
@@ -128,5 +128,5 @@ Atmospheric_inputs:
 Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
-  Photodetector noise: False
+  Photodetector noise: True
   Save Scanning Coordinates: False
