### 2021-04-21 #3109 @danielskatz
@PacoCosta - we have a few things we need to do now
1. Get the paper to compile - I'll check and see if you have made changes. You can also do this, by using the command `@whedon check references` to check the references, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
2. Add an explicit "Statement of Need" section
3. suggest some additional reviewers - if you have any suggestions then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
