+++ Docs/JOSS_202102/paper.bib
@@ -109,7 +109,7 @@
   	url = {https://journals.ametsoc.org/view/journals/apme/7/1/1520-0450_1968_007_0105_tdokpo_2_0_co_2.xml},
   	Archiveprefix = {},
   	Author = {Browning, K.A. and Wexler, R.},
-  	DOI = {https://doi.org/10.1175/1520-0450(1968)007<0105:TDOKPO>2.0.CO;2},
+  	DOI = {10.1175/1520-0450(1968)007<0105:TDOKPO>2.0.CO;2},
   	Journal = {{Journal of Applied Meteorology and Climatology}},  	
   	Title = {{The Determination of Kinematic Properties of a Wind Field Using Doppler Radar}},
    License = {{Full Access}},
