### 2021-04-21 #3109 @danielskatz
Re the reference, I see "Costa, F. (2021). Qlunc: A Python framework for quantification of lidar uncertainty. In Journal of Open Source Software (Version 0.91)." This seems to be a reference to this paper in JOSS, not the software in Zenodo.  But in either case, since the software will be published as part of the paper, why would the paper either cite itself or an element of it?
