+++ Docs/JOSS_202102/paper.bib
@@ -50,18 +50,6 @@
   url          = {https://doi.org/10.5281/zenodo.4432136}
 }
 
-@misc{zenodo,
-  	url = {},
-  	Archiveprefix = {},
-  	Author = {Costa, F.},
-  	DOI = {},
-  	Title = {{Qlunc: A Python framework for quantification of lidar uncertainty}},
-    	Abstract = {Numerical framework to assess lidar hardware and data processing methods},
-     Journal={{Journal of Open Source Software}},
-    	License = {BSD 3-Clause "New" or "Revised" License},
-    	Version = {0.91},
-  	Year = 2021
-}
 
 @software{yaddum,
   author       = {Vasiljevic, Nikola},
