### 2021-04-22 #3109 @danielskatz
👋 @NickLaurenson, @PierreGuilbertF, @bastienjacquet - would one of you be willing to review this JOSS submission (short paper and software)?  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
