+++ Tutorials/WorkingExample_Tutorial.py
@@ -118,7 +118,7 @@ class optical_amplifier():
     def __init__(self,name,OA_NF,OA_Gain,unc_func):
                  self.Optical_AmplifierID = name
                  self.NoiseFig            = OA_NF
-                 self.Gain                = OA_Gain
+                 self.OA_Gain                = OA_Gain
                  self.Uncertainty         = unc_func
                  print('Created new optical amplifier: {}'.format(self.Optical_AmplifierID))
         
