+++ Tutorials/Qlunc_Instantiate_Tutorial.py
@@ -1,0 +1,150 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Oct 19 11:08:32 2020
+@author: fcosta
+
+Francisco Costa García
+University of Stuttgart(c) 
+
+###############################################################################
+###################### __ Instantiate __ ######################################
+
+Digitally creating a virtual object which represents a physical, real object
+could be done by instantiating their python classes:
+        
+        1) Run Qlunc_Classes.py to create the classes corresponding to
+           components, modules, atmospheric scenarios and lidar device
+        2) Instantiate the different components
+        3) Instantiate modules including corresponding components
+        4) Instantiate class `atmosphere` --> atmospheric conditions
+        5) Instantiate class `lidar_gral_inp` --> lidar general inputs
+        6) Instantiate class `lidar` by including modules, lidar general inputs 
+           and atmospheric conditions
+           
+"""
+import os
+# os.chdir('../')
+# importing  uncertainty functions
+pdb.set_trace()
+import UQ_Functions.UQ_Photonics_Classes as uphc,UQ_Functions.UQ_Optics_Classes as uopc, UQ_Functions.UQ_Power_Classes as upwc,UQ_Functions.UQ_Lidar_Classes as ulc
+from Utils.Qlunc_ImportModules import *
+
+#%% Running Qlunc_Classes.py:
+with open (r'./yaml_inputs_file1.yml') as file: # WHere the yaml file is in order to get the input data
+    Qlunc_yaml_inputs={}
+    docs = yaml.load_all(file, Loader=yaml.FullLoader)
+    for doc in docs:      
+        for k, v in doc.items():           
+            Qlunc_yaml_inputs.setdefault(k,v)  # save a dictionary with the data coming from yaml file 
+
+# Execute Qlunc_Classes.py (creating classes for lidar 'objects')
+exec(open(Qlunc_yaml_inputs['Main_directory']+'/Tutorials/WorkingExample_Tutorial.py').read())   
+
+#%%%%%%%%%%%%%%%%% INPUTS %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+## Optics components and Module: ##############################################
+
+# Here we create optics components and optics module. User can create as many components as he/she want and combine them to create different module types
+# Each module/component is a python object with their own technical characteristics and can be flexible combined to assess different use cases. 
+
+# Scanner:
+Scanner           = scanner(name            = Qlunc_yaml_inputs['Components']['Scanner']['Name'],           # Introduce your scanner name.
+                            scanner_type    = Qlunc_yaml_inputs['Components']['Scanner']['Type'],
+                            origin          = Qlunc_yaml_inputs['Components']['Scanner']['Origin'],         # Origin (coordinates of the lidar deployment).
+                            pattern         = Qlunc_yaml_inputs['Components']['Scanner']['Pattern'],
+                            sample_rate     = Qlunc_yaml_inputs['Components']['Scanner']['Sample rate'],    # for now introduce it in [degrees].
+                           
+                           # This values for focus distance, cone_angle and azimuth define a typical VAD scanning sequence:
+                            azimuth         = np.array(np.arange(Qlunc_yaml_inputs['Components']['Scanner']['Azimuth'][0],                                                  
+                                                                 Qlunc_yaml_inputs['Components']['Scanner']['Azimuth'][1],
+                                                                 Qlunc_yaml_inputs['Components']['Scanner']['Azimuth'][2])), # Azimuth angle in [degrees].
[…]
