+++ Tutorials/Qlunc_Instantiate_Tutorial.py
@@ -23,9 +23,9 @@ could be done by instantiating their python classes:
            
 """
 import os
-
+print(os.getcwd())
 # importing  uncertainty functions
-
+# os.chdir('../')
 import UQ_Functions.UQ_Photonics_Classes as uphc,UQ_Functions.UQ_Optics_Classes as uopc, UQ_Functions.UQ_Power_Classes as upwc,UQ_Functions.UQ_Lidar_Classes as ulc
 from Utils.Qlunc_ImportModules import *
 print(os.getcwd())
