### 2021-05-04 #3211 @PacoCosta
Fixing Figure1. Qlunc basic structure. Power module components were wrong. No `issue` regarding this.
