### 2021-05-07 afa066a
Including laser
Including laser  in the photonics module
