### 2021-05-07 b5cbb20
Optical circulator update
I assume an SNR for noise calculations. Also laser is included (function calculating its uncertainty is not implemented yet. Instead, a temporal solution is implemented
