### 2021-05-07 ef3a76d
Optical circulator
Assume an SNR closes #107
