### 2021-05-07 e495c25
Update UQ_Photonics_Classes.py
Laser comments
