### 2021-05-08 7028e74
Include telescope & laser classes
