### 2021-05-24 #3211 @adi3
@PacoCosta Can you please edit your paper to write Qlunc in regular font instead of as a code block?
In addition, can this be installed via `pip` or `conda`? I don't see it mentioned in the installation instructions.
