### 2021-05-24 #3211 @adi3
@PacoCosta A few more points.

1. Please remove the `Usage` and `Tutorials` section from your paper. These do not go in a JOSS paper.
2. Please combine the `Summary` and `Qlun's capabilities` section. As it is, the `Summary` section a is bit too verbose right now and I would recommend you in condensing it. Thereafter, add the content from `Qlun's capabilities` under `Summary` and ensure there is no repetition.
3. The last line of your current `Summary` section is not needed. Please remove it. The JOSS paper will have a link to your repository anyway.
