+++ Docs/JOSS_202102/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 Wind lidar is a flexible and versatile remote sensing device for wind energy applications [@Hauke] which measures the wind vector remotely using laser light backscattered from aerosols. They are a key tool for wind energy and meteorology. Like any measurement method, it is essential to estimate their uncertainty.
 Qlunc, which stands for **Q**uantification of **l**idar **unc**ertainty, is an open-source python-based tool to create a digital twin of the lidar hardware, and estimate the uncertainty of wind lidar wind speed measurements.
 
-Qlunc contains models of the uncertainty contributed by individual lidar components and modules (represented by python objects, which in turn represent physical lidar  objects), that are then easily combined (thanks to the modularity of the code) considering their different nature, to estimate the uncertainties in wind lidar measurements. The modules are based on the OpenLidar architecture [@OpenLidar] and can be adapted for particular use cases (see \autoref{fig:QluncStructure}). The terminology for the components and modules defined within Qlunc has also been aligned with a community-driven wind lidar ontology, which is in development [@OntoStack;@sheet2rdf]. 
+Qlunc contains models of the uncertainty contributed by individual lidar components and modules (represented by python objects, which in turn represent physical lidar  objects), that are then easily combined considering their different nature, to estimate the uncertainties in wind lidar measurements. The modules are based on the OpenLidar architecture [@OpenLidar] and can be adapted for particular use cases thanks to the modularity of the code (see \autoref{fig:QluncStructure}). The terminology for the components and modules defined within Qlunc has also been aligned with a community-driven wind lidar ontology, which is in development [@OntoStack;@sheet2rdf]. 
 
 ![Qlunc basic structure.\label{fig:QluncStructure}](Qlunc_BasicStructure_diagram.png)
  
