### 2021-05-30 d0cb8a1
Lissajous parameters
Including inputs for lissajous pattern in the yaml file - closes #122
