+++ Main/Qlunc_inputs.yml
@@ -18,13 +18,15 @@ Components:
   Scanner:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
-    Origin: [0,0,0]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
+    Origin: [0,0,60]  # Lidar position in meters: [x,y,z]    
+    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
+    # Lissajous patterns:
+    Lissajous parameters: [50,60,60,3,3] # [x_dim,y_dim,z_dim,q,p]
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
+    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
   
     # STDV in lidar angles and focus distance
-    stdv focus distance: 100   # [m]
+    stdv focus distance: 1   # [m]
     stdv Cone angle: .2      # [deg]
     stdv Azimuth: 0         # [deg]   
    
@@ -91,7 +93,7 @@ Components:
 # Lidar general inputs 
   Lidar general inputs:
     Name: Lidar_gral_inputs_Yaml
-    Type: Pulsed # CW/Pulsed for continuous or pulsed lidars
+    Type: CW # CW/Pulsed for continuous or pulsed lidars
     Wavelength: 1550.e-9    # [m]
     Yaw error: 0            # [deg]
     Pitch error: 0          # [deg]
@@ -140,5 +142,5 @@ Atmospheric_inputs:
 Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
-  Photodetector noise: True
+  Photodetector noise: False
   Save Scanning Coordinates: False
