### 2021-05-30 3b62cf0
implement AOM
main structure of AOM implemented as losses in AOM
