+++ Main/Qlunc_inputs.yml
@@ -19,11 +19,11 @@ Components:
     
     Name: Scanner_Yaml # Provide an ID to your lidar:    
     Origin: [0,0,60]  # Lidar position in meters: [x,y,z]    
-    Pattern: 'lissajous' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
+    Pattern: 'None' # 'lissajous', user defined pattern or 'None' in case none predefined pattern is used
     # Lissajous patterns:
     Lissajous parameters: [50,60,60,3,3] # [x_dim,y_dim,z_dim,q,p]
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
+    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
   
     # STDV in lidar angles and focus distance
     stdv focus distance: .01   # [m]
@@ -31,9 +31,9 @@ Components:
     stdv Azimuth: 0.5         # [deg]   
    
     # VAD Pattern:
-    Focus distance: [70]   # [m]
+    Focus distance: [20]   # [m]
     Cone angle: [30]       # [deg]
-    Azimuth: [0,360,45]     # [deg] [origin,end,steps]
+    Azimuth: [0,360,360]     # [deg] [origin,end,steps]
 
     # Scanning Lidar Pattern:
     x: [-40]
@@ -86,7 +86,7 @@ Components:
   Laser:
     Name: Laser_YAML
     Wavelength: 1550.e-9    # [m]
-    Stdv Wavelength: 1.e-12 #[m]
+    Stdv Wavelength: 1.e-9 #[m]
     Output power: 1.e-6 #[w]
     Uncertainty function: uphc.UQ_Laser
 
@@ -99,9 +99,9 @@ Components:
     Name: Lidar_gral_inputs_Yaml
     Type: CW # CW/Pulsed for continuous or pulsed lidars. The difference is that AOM (photonics module) is included in calculations when Pulsed
     Wavelength: 1550.e-9    # [m]
-    Yaw error: 0            # [deg]
-    Pitch error: 0          # [deg]
-    Roll error: 0          # [deg]
+    Yaw error: 9            # [deg]
+    Pitch error: 0         # [deg]
+    Roll error: 2          # [deg]
 ---
 # Modules:
 Modules:
