+++ Main/Qlunc_inputs.yml
@@ -122,7 +122,7 @@ Modules:
 Lidar:
   Name: Caixa1
   Photonics module: Photonics_Module       # Have to be the same name as the instance
-  Optics module: Optics_Module             # Have to be the same name as the instance
+  Optics module: None #Optics_Module             # Have to be the same name as the instance
   Power module: None                       # Have to be the same name as the instance
   Lidar inputs: Lidar_inputs               # Have to be the same name as the instance
   Uncertainty function: ulc.sum_unc_lidar  # Python module! (You develop a function and save it as a python module)
@@ -145,6 +145,6 @@ Atmospheric_inputs:
 # Plotting:------------------------------------------------------------
 Flags:
   Pointing accuracy uncertainty: False
-  Scanning Pattern: True
+  Scanning Pattern: False
   Photodetector noise: False
   Save Scanning Coordinates: False
