### 2021-06-11 280d262
Testing photonics
Testing 'module photonics = None'.  Issue with the optical circulator.
When module photonics = None --> optical circulator seems to fail
