+++ readme.md
@@ -75,6 +75,7 @@ The following python libraries and tools should be installed beforehand and are 
 - scipy==1.6.0
 - sympy==1.7.1
 - xarray==0.15.1
+- xarray-extras==0.4.2
 - python==2.7
 - spyder==4.2.1
 - netcdf4
