+++ readme.md
@@ -76,6 +76,8 @@ The following python libraries and tools should be installed beforehand and are 
 - sympy==1.7.1
 - xarray==0.15.1
 - xarray-extras==0.4.2
+- numba
+- dask
 - python==2.7
 - spyder==4.2.1
 - netcdf4
