### 2021-06-17 250fe99
Xarray
Implementing code to allow outputs to fit into mocalum
