+++ UQ_Functions/UQ_Optics_Classes.py
@@ -185,7 +185,7 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
 def UQ_OpticalCirculator(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     """
     Optical circulator uncertainty estimation. Location: ./UQ_Functions/UQ_Optics_Classes.py
-    It assumes an SNR and calculates the Optical circulator noise from it: $SNR\,=\,\frac{{P_{i}},{P_{noise}}$
+    
     Parameters
     ----------
     
@@ -204,12 +204,19 @@ def UQ_OpticalCirculator(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     list
     
     """
-    # Optical_Circulator_Uncertainty = [np.array(Lidar.optics.optical_circulator.insertion_loss)]
-    Optical_Circulator_Uncertainty_w = [Qlunc_yaml_inputs['Components']['Laser']['Output power']/(10**(Lidar.optics.optical_circulator.SNR/10))]
-    Optical_Circulator_Uncertainty_dB = 10*np.log10(Optical_Circulator_Uncertainty_w)
-    Final_Output_UQ_Optical_Circulator={'Optical_Circulator_Uncertainty':Optical_Circulator_Uncertainty_dB}
-    Lidar.lidar_inputs.dataframe['Optical circulator']=Final_Output_UQ_Optical_Circulator['Optical_Circulator_Uncertainty']
-
+    
+    
+    Optical_Circulator_losses = [np.array(Lidar.optics.optical_circulator.insertion_loss)]
+    
+    #  if the insertion loss is expressed in % (X% losses):    
+    #Optical_Circulator_losses = 10*np.log10(1-(X/100)) # output in dB
+    
+    # # If it is expressed using SNR:
+    # Optical_Circulator_losses = [Qlunc_yaml_inputs['Components']['Laser']['Output power']*(10**(Lidar.optics.optical_circulator.SNR/10))]
+    # Optical_Circulator_Uncertainty_dB = 10*np.log10(Optical_Circulator_Uncertainty_w)
+    # Final_Output_UQ_Optical_Circulator={'Optical_Circulator_Uncertainty':Optical_Circulator_Uncertainty_dB}
+    # Lidar.lidar_inputs.dataframe['Optical circulator']=Final_Output_UQ_Optical_Circulator['Optical_Circulator_Uncertainty']
+ 
     return Final_Output_UQ_Optical_Circulator,Lidar.lidar_inputs.dataframe
 
 #%% TELESCOPE NOT IMPLEMENTED
