+++ UQ_Functions/UQ_Lidar_Classes.py
@@ -91,11 +91,9 @@ def sum_unc_lidar(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
         names     = [Lidar.LidarID]
         component = [i for i in DataXarray.keys()]
         data      = [ii for ii in DataXarray.values()]
-        
         df = xr.DataArray(data,
                 coords = [component,names],
                 dims   = ('Components','Names'))
-        pdb.set_trace()
         if not os.path.exists('./Projects'):
             os.makedirs('./Projects')
         
