### 2021-06-27 #3211 @adi3
@PacoCosta I'm almost finished with my review. I see only one item still outstanding. Could you please community guidelines for third parties wishing to
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

This can either be done by adding a CONTRIBUTING.md file and linking it in the README, or adding the content directly in the main repo README. Thanks!
