### 2021-06-27 385e910
adding name of the project to yaml file
