+++ UQ_Functions/UQ_Optics_Classes.py
@@ -56,16 +56,14 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     sample_rate_count=0
                 
     # R: Implement error in deployment of the tripod as a rotation over yaw, pitch and roll
-    stdv_yaw    = np.deg2rad(Lidar.lidar_inputs.yaw_error_dep)
-    stdv_pitch  = np.deg2rad(Lidar.lidar_inputs.pitch_error_dep)
-    stdv_roll   = np.deg2rad(Lidar.lidar_inputs.roll_error_dep)
-    
-    R=[[np.cos(stdv_yaw)*np.cos(stdv_pitch) ,  np.cos(stdv_yaw)*np.sin(stdv_pitch)*np.sin(stdv_roll)-np.sin(stdv_yaw)*np.cos(stdv_roll) ,  np.cos(stdv_yaw)*np.sin(stdv_pitch)*np.cos(stdv_roll)+np.sin(stdv_yaw)*np.sin(stdv_roll)],
-      [np.sin(stdv_yaw)*np.cos(stdv_pitch)  ,  np.sin(stdv_yaw)*np.sin(stdv_pitch)*np.sin(stdv_roll)+np.cos(stdv_yaw)*np.cos(stdv_roll) ,  np.sin(stdv_yaw)*np.sin(stdv_pitch)*np.cos(stdv_roll)-np.cos(stdv_yaw)*np.sin(stdv_roll)],
-      [       -np.sin(stdv_pitch)           ,  np.cos(stdv_pitch)*np.sin(stdv_roll)                                                     ,  np.cos(stdv_pitch)*np.cos(stdv_roll)]]
+    stdv_yaw    = np.array(np.deg2rad(Lidar.lidar_inputs.yaw_error_dep))
+    stdv_pitch  = np.array(np.deg2rad(Lidar.lidar_inputs.pitch_error_dep))
+    stdv_roll   = np.array(np.deg2rad(Lidar.lidar_inputs.roll_error_dep))
     
     
+
     
+    # stdv focus distance, cone angle and azimuth:
     stdv_param1 = Lidar.optics.scanner.stdv_focus_dist    
     stdv_param2 = np.deg2rad(Lidar.optics.scanner.stdv_cone_angle)
     stdv_param3 = np.deg2rad(Lidar.optics.scanner.stdv_azimuth)
@@ -111,8 +109,8 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
                 param3.append((np.pi)+(np.arctan(y_init[ind]/x_init[ind])))            
             elif x_init[ind]==0:
                 param3.append(np.pi/2.0*(np.sign(y_init[ind])))
-   
-    for param1_or,param2_or,param3_or in zip(param1,param2,param3):# Take coordinates from inputs
+    # pdb.set_trace()
+    for param1_or,param2_or,param3_or,stdv_yaw_or,stdv_pitch_or, stdv_roll_or in zip(param1,param2,param3,stdv_yaw,stdv_pitch,stdv_roll):# Take coordinates from inputs
         Mean_DISTANCE=[]
         DISTANCE=[]        
         stdv_DISTANCE=[]  
@@ -126,25 +124,38 @@ def UQ_Scanner(Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
         X0.append(x0)
         Y0.append(y0)
         Z0.append(z0)
+        # ind_incl = 0 
         for trial in range(0,10):
             
             # Create white noise with stdv selected by user:
-            n=10000 # Number of cases to combine
+            n=100 # Number of cases to combine           
+            # Position, due to pointing accuracy
[…]
