+++ Utils/Qlunc_Help_standAlone.py
@@ -16,7 +16,7 @@ flatten = lambda *n: (e for a in n for e in (flatten(*a) if isinstance(a, (list,
 #%% sum dB:
 def sum_mat(noisy_yaw,noisy_pitch, noisy_roll):
     R=[]
-    for i in range(3):
+    for i in range(len(noisy_yaw)):
     # R.append([[np.cos(noisy_yaw)*np.cos(noisy_pitch) ,  np.cos(noisy_yaw)*np.sin(noisy_pitch)*np.sin(noisy_roll)-np.sin(noisy_yaw)*np.cos(noisy_roll) ,  np.cos(noisy_yaw)*np.sin(noisy_pitch)*np.cos(noisy_roll)+np.sin(noisy_yaw)*np.sin(noisy_roll)],
     #            [np.sin(noisy_yaw)*np.cos(noisy_pitch)  ,  np.sin(noisy_yaw)*np.sin(noisy_pitch)*np.sin(noisy_roll)+np.cos(noisy_yaw)*np.cos(noisy_roll) ,  np.sin(noisy_yaw)*np.sin(noisy_pitch)*np.cos(noisy_roll)-np.cos(noisy_yaw)*np.sin(noisy_roll)],
     #            [       -np.sin(noisy_pitch)           ,  np.cos(noisy_pitch)*np.sin(noisy_roll)                                                     ,  np.cos(noisy_pitch)*np.cos(noisy_roll)]])
@@ -25,7 +25,7 @@ def sum_mat(noisy_yaw,noisy_pitch, noisy_roll):
                   [       -np.sin(noisy_pitch[i])           ,  np.cos(noisy_pitch[i])*np.sin(noisy_roll[i])                                                     ,  np.cos(noisy_pitch[i])*np.cos(noisy_roll[i])]])
     R_mean=np.sum(R,axis=0)/len(noisy_yaw)
     # pdb.set_trace()
-    return np.degrees(R_mean)
+    return R_mean
 def sum_dB(data,uncorrelated):
     """
     Add up dB's. Location: Qlunc_Help_standAlone.py
