+++ Main/Qlunc_inputs.yml
@@ -26,9 +26,9 @@ Components:
     Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
   
     # STDV in lidar angles and focus distance
-    stdv focus distance: .01   # [m]
-    stdv Cone angle: .25      # [deg]
-    stdv Azimuth: 0.5         # [deg]   
+    stdv focus distance: 0   # [m]
+    stdv Cone angle: 0      # [deg]
+    stdv Azimuth: 0         # [deg]   
    
     # VAD Pattern:
     Focus distance: [20]   # [m]
@@ -102,9 +102,9 @@ Components:
     Name: Lidar_gral_inputs_Yaml
     Type: CW # CW/Pulsed for continuous or pulsed lidars. The difference is that AOM (photonics module) is included in calculations when Pulsed
     Wavelength: 1550.e-9    # [m]
-    Yaw error: [2]       # [deg]
-    Pitch error: [0]         # [deg]
-    Roll error: [0]          # [deg]
+    Yaw error: 0       # [deg]
+    Pitch error: 0         # [deg]
+    Roll error: 0          # [deg]
 ---
 # Modules:
 Modules:
