### 2021-06-27 a14c98b
R mat scanner
It works except when Time series = True
