+++ UQ_Functions/UQ_Lidar_Classes.py
@@ -42,65 +42,65 @@ def sum_unc_lidar(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     try: # ecah try/except evaluates wether the component is included in the module, therefore in the calculations
 #        if Photodetector_Uncertainty not in locals():
         # pdb.set_trace()
-        Photonics_Uncertainty,DataFrame=Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+        Photonics_Uncertainty,DataFrame = Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
         List_Unc_lidar.append(Photonics_Uncertainty['Uncertainty_Photonics'])
     except:
-        Photonics_Uncertainty=None
+        Photonics_Uncertainty = None
         print('No photonics module in calculations!')
     try:
-        Optics_Uncertainty,DataFrame=Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
-        Optics_Uncertainty=np.ndarray.tolist(Optics_Uncertainty['Uncertainty_Optics'])*len(Atmospheric_Scenario.temperature)
+        Optics_Uncertainty,DataFrame = Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+        Optics_Uncertainty           = np.ndarray.tolist(Optics_Uncertainty['Uncertainty_Optics'])*len(Atmospheric_Scenario.temperature)
         List_Unc_lidar.append(np.array([Optics_Uncertainty]))
     except:
-        Optics_Uncertainty=None
+        Optics_Uncertainty = None
         print('No optics module in calculations!')
     try:
-        Power_Uncertainty,DataFrame=Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
+        Power_Uncertainty,DataFrame = Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
         List_Unc_lidar.append(Power_Uncertainty['Uncertainty_Power']*len(Atmospheric_Scenario.temperature))
     except:
-        Power_Uncertainty=None
+        Power_Uncertainty = None
         print('No power module in calculations!')
     print('Processing lidar uncertainties...')
-    Uncertainty_Lidar=SA.unc_comb(List_Unc_lidar)
-    Final_Output_Lidar_Uncertainty = {'Lidar_Uncertainty':Uncertainty_Lidar}    
-    Lidar.lidar_inputs.dataframe['Lidar']=Final_Output_Lidar_Uncertainty['Lidar_Uncertainty']
+    Uncertainty_Lidar                     = SA.unc_comb(List_Unc_lidar)
+    Final_Output_Lidar_Uncertainty        = {'Lidar_Uncertainty':Uncertainty_Lidar}    
+    Lidar.lidar_inputs.dataframe['Lidar'] = Final_Output_Lidar_Uncertainty['Lidar_Uncertainty']
     
     ########################################################################################################
     # Create Xarray to store data. Link with Mocalum and yaddum  ###########################################
     
-    # DataXarray=Lidar.lidar_inputs.dataframe
+    DataXarray=Lidar.lidar_inputs.dataframe
 
-    # if os.path.isfile('./Projects/' + Qlunc_yaml_inputs['Project']+ '.nc'):
-    #     # Read the new lidar data
-    #     names     = [Lidar.LidarID]
-    #     component = [i for i in DataXarray.keys()]
-    #     data      = [ii for ii in DataXarray.values()]
[…]
