### 2021-06-27 fb3c6ae
Update UQ_Optics_Classes.py
closes #134
