+++ environment.yml
@@ -16,6 +16,7 @@ dependencies:
   - sympy==1.7.1
   - xarray==0.15.1
   - spyder==4.2.1
-  - python
+  - python==3.7.9
   - notebook
   - jupyterlab
+  - netcdf
