### 2021-06-28 6a9e660
Update Qlunc_Classes.py
Include 'time' in the class atmosphere
