+++ readme.md
@@ -85,6 +85,11 @@ The following python libraries and tools should be installed beforehand and are 
 ## Author
 [Francisco Costa](https://www.ifb.uni-stuttgart.de/en/institute/team/Costa-Garcia/)
 
+## Contributions:
+If you are wishing to:
+- Colaborate: please contact the author
+- Report issues: follow [Github guidelines to report code issues](https://docs.github.com/en/issues/tracking-your-work-with-issues/quickstart)
+- Seek for support: please contact the author
 ## License
 `Qlunc` is licensed under **[SD 3-Clause License](https://github.com/SWE-UniStuttgart/Qlunc/blob/main/LICENSE)**
 
