### 2021-06-28 #3211 @PacoCosta
https://github.com/openjournals/joss-reviews/issues/3211#issuecomment-869146287: Done
Thanks.
