+++ Main/Qlunc_inputs.yml
@@ -23,7 +23,7 @@ Components:
     # Lissajous patterns:
     Lissajous parameters: [50,60,60,3,3] # [x_dim,y_dim,z_dim,q,p]
     Sample rate: 0   # Lidar sample rate. In meters. Use 0 for now.
-    Type: VAD    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
+    Type: SCAN    # Type of the scanning head: VAD or SCAN (Scanning Lidar).
   
     # STDV in lidar angles and focus distance
     stdv focus distance: 0   # [m]
@@ -103,8 +103,8 @@ Components:
     Type: CW # CW/Pulsed for continuous or pulsed lidars. The difference is that AOM (photonics module) is included in calculations when Pulsed
     Wavelength: 1550.e-9    # [m]
     Yaw error: 0       # [deg]
-    Pitch error: 0         # [deg]
-    Roll error: 0          # [deg]
+    Pitch error: 10         # [deg]
+    Roll error: 15          # [deg]
 ---
 # Modules:
 Modules:
@@ -149,5 +149,5 @@ Atmospheric_inputs:
 Flags:
   Pointing accuracy uncertainty: False
   Scanning Pattern: True
-  Photodetector noise: False
+  Photodetector noise: True
   Save Scanning Coordinates: False
