### 2021-07-01 de7a9a9
Checking ASE in optical amplifier
