+++ Main/Qlunc_Classes.py
@@ -117,10 +117,11 @@ class photodetector():
                  print('Created new photodetector: {}'.format(self.PhotodetectorID))
         
 class optical_amplifier():
-    def __init__(self,name,NoiseFig,OA_Gain,unc_func):
+    def __init__(self,name,NoiseFig,OA_Gain,OA_BW,unc_func):
                  self.Optical_AmplifierID = name
                  self.NoiseFig            = NoiseFig
                  self.OA_Gain             = OA_Gain
+                 self.OA_BW               = OA_BW
                  self.Uncertainty         = unc_func
                  print('Created new optical amplifier: {}'.format(self.Optical_AmplifierID))
 
