### 2021-07-02 f87b01e
Update UQ_Optics_Classes.py
Option to include the different components
