### 2021-07-02 6406dd0
Update UQ_Photonics_Classes.py
Optical amplifier with actual data
