+++ UQ_Functions/UQ_Lidar_Classes.py
@@ -39,27 +39,38 @@ def sum_unc_lidar(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
     
     """ 
     List_Unc_lidar = []
-    try: # ecah try/except evaluates wether the component is included in the module, therefore in the calculations
-#        if Photodetector_Uncertainty not in locals():
-        # pdb.set_trace()
-        Photonics_Uncertainty,DataFrame = Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
-        List_Unc_lidar.append(Photonics_Uncertainty['Uncertainty_Photonics'])
-    except:
-        Photonics_Uncertainty = None
-        print('No photonics module in calculations!')
-    try:
-        Optics_Uncertainty,DataFrame = Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
-        Optics_Uncertainty           = np.ndarray.tolist(Optics_Uncertainty['Uncertainty_Optics'])*len(Atmospheric_Scenario.temperature)
-        List_Unc_lidar.append(np.array([Optics_Uncertainty]))
-    except:
-        Optics_Uncertainty = None
-        print('No optics module in calculations!')
-    try:
-        Power_Uncertainty,DataFrame = Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
-        List_Unc_lidar.append(Power_Uncertainty['Uncertainty_Power']*len(Atmospheric_Scenario.temperature))
-    except:
-        Power_Uncertainty = None
-        print('No power module in calculations!')
+    
+    if Lidar.photonics != 'None':
+        try: # ecah try/except evaluates wether the component is included in the module, therefore in the calculations
+    #        if Photodetector_Uncertainty not in locals():
+            # pdb.set_trace()
+            Photonics_Uncertainty,DataFrame = Lidar.photonics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+            List_Unc_lidar.append(Photonics_Uncertainty['Uncertainty_Photonics'])
+        except:
+            Photonics_Uncertainty = None
+            print('Error in photonics module calculations!')
+    else:
+        print('You didn´t include a photonics module in the lidar')
+    if Lidar.optics != 'None':
+        try:
+            Optics_Uncertainty,DataFrame = Lidar.optics.Uncertainty(Lidar,Atmospheric_Scenario,cts,Qlunc_yaml_inputs)
+            Optics_Uncertainty           = np.ndarray.tolist(Optics_Uncertainty['Uncertainty_Optics'])*len(Atmospheric_Scenario.temperature)
+            List_Unc_lidar.append(np.array([Optics_Uncertainty]))
+        except:
+            Optics_Uncertainty = None
+            print('Error in optics module calculations!')
+    else:
+        print('You didn´t include an optics module in the lidar')
+    if Lidar.power !=0:
+        
+        try:
+            Power_Uncertainty,DataFrame = Lidar.power.Uncertainty(Lidar,Atmospheric_Scenario,cts)
+            List_Unc_lidar.append(Power_Uncertainty['Uncertainty_Power']*len(Atmospheric_Scenario.temperature))
+        except:
+            Power_Uncertainty = None
+            print('No power module in calculations!')
+    else:
+        print('You didn´t include a power module in  the lidar')
     print('Processing lidar uncertainties...')
     Uncertainty_Lidar                     = SA.unc_comb(List_Unc_lidar)[0]
     # pdb.set_trace()
