### 2021-07-20 f2e0061
introducing probe volume in calculations
- class
- instance
- zr calculation
- zr output for vicondar as a .txt
