+++ UQ_Functions/UQ_ProbeVolume_Classes.py
@@ -1,0 +1,112 @@
+# # -*- coding: utf-8 -*-
+# """
+# Created on Wed May 19 12:48:10 2021
+
+# @author: fcosta
+# """
+# '''
+# This works for a CW monostatic coherent lidar 
+# '''
+
+
+from Utils.Qlunc_ImportModules import *
+from Utils import Qlunc_Help_standAlone as SA
+from Utils import Scanning_patterns as SP
+from Utils import Qlunc_Plotting as QPlot
+
+def UQ_Probe_volume (Lidar, Atmospheric_Scenario,cts,Qlunc_yaml_inputs):
+    
+    # f_length  = 200e-3 # focal length
+    # Qlunc_yaml_inputs['Probe Volume']['Fiber-lense distance']         = np.arange(2e-3,4e-3,.02e-3) # distance fiber-end--telescope lens
+    # a0        = 198.1e-3 # the offset (a constant number), to avoid the fiber-end locates at the focal point, otherwise the lights will be parallel to each other
+    # A         = 20e-3 # beam radius at the output lens
+    # ext_coef  = 0.085
+    # effective_radius_telescope  = 16.5e-3
+    
+    # %% Liqin jin
+    if Qlunc_yaml_inputs['Components']['Lidar general inputs']['Type']=="CW":
+        # The focus distance varies with the distance between the fiber-end and the telescope lens. So that, also the probe length varies with such distance.
+        #Calculating focus distance depending on the distance between the fiber-end and the telescope lens:
+        f_distance = 1/((1/Qlunc_yaml_inputs['Probe Volume']['Focal length'])-(1/(Qlunc_yaml_inputs['Probe Volume']['Fiber-lens distance']+Qlunc_yaml_inputs['Probe Volume']['Fiber-lens offset']))) # Focus distance
+        # dist =(np.linspace(0,60,len(Qlunc_yaml_inputs['Probe Volume']['Fiber-lens distance'])))  # distance from the focus position along the beam direction
+        
+        # Rayleigh length variation due to f_distance variations (due to the distance between fiber-end and telescope lens)
+        zr = (Qlunc_yaml_inputs['Components']['Laser']['Wavelength']*(f_distance**2))/(np.pi*(Qlunc_yaml_inputs['Probe Volume']['Effective radius telescope'])**2)# Rayleigh length  (considered as the probe length) # half-width of the weighting function --> FWHM = 2*zr
+        
+        # Saving coordenates to a file in desktop
+        if os.path.isfile('./metadata/rayleigh_distance.txt'):
+            os.remove('./metadata/rayleigh_distance.txt')
+            file=open('./metadata/rayleigh_distance.txt','w')
+            file.write(repr(zr))
+            file.close()   
+            pdb.set_trace()
+        else:
+            file=open('./metadata/rayleigh_distance.txt','w')
+            file.write(repr(zr))
+            file.close() 
+        # Probe volume:
+        #Probe_volume = np.pi*(Qlunc_yaml_inputs['Probe Volume']['Output beam radius']**2)*((4*(f_distance**2)*Qlunc_yaml_inputs['Components']['Laser']['Wavelength'])/(Telescope_aperture)) # based on Marijn notes
+        #VolCil       = np.pi*(Qlunc_yaml_inputs['Probe Volume']['Output beam radius']**2)*fwhm  # calculated based on the fwhm
[…]
