+++ metadata/rayleigh_distance.txt
@@ -1,0 +1,1 @@
+286.76082290809364

[…]
