### 2021-07-21 b15b74b
Uncertainty in focus distance
Due to variations in distance between fiber-end and telescope lens the focus distance varies. The uncertainty in this distance varies as well with focal length, offset distance (a0) and fiber-lens distance(a)
