+++ metadata/rayleigh_distance.txt
@@ -1,1 +1,1 @@
-286.76082290809364
+14.238114853496493
