+++ Main/Qlunc_Classes.py
@@ -191,25 +191,27 @@ class optical_circulator():
                  print ('Created new optical circulator: {}'.format(self.Optical_CirculatorID))
 
 class telescope():
-    def __init__(self,name,aperture,stdv_aperture,unc_func):
-                self.TelescopeID   = name
-                self.aperture      = aperture
-                self.stdv_aperture = stdv_aperture
-                self.Uncertainty   = unc_func
+    def __init__(self,name,aperture,stdv_aperture,focal_length,fiber_lens_d,fiber_lens_offset,effective_radius_telescope,output_beam_radius,stdv_fiber_lens_d,stdv_fiber_lens_offset,stdv_focal_length,stdv_eff_radius_telescope ,unc_func):
+                self.TelescopeID                = name
+                self.aperture                   = aperture
+                self.stdv_aperture              = stdv_aperture
+                self.focal_length               = focal_length
+                self.fiber_lens_d               = fiber_lens_d
+                self.fiber_lens_offset          = fiber_lens_offset
+                self.effective_radius_telescope = effective_radius_telescope
+
+                self.output_beam_radius         = output_beam_radius
+                self.stdv_fiber_lens_d          = stdv_fiber_lens_d
+                self.stdv_fiber_lens_offset     = stdv_fiber_lens_offset
+                self.stdv_focal_length          = stdv_focal_length
+                self.stdv_eff_radius_telescope  = stdv_eff_radius_telescope 
+                self.Uncertainty                = unc_func
                 print('Created new telescope: {}'.format(self.TelescopeID))
 
 class probe_volume():
-    def __init__(self,name,focal_length,fiber_lens_d,fiber_lens_offset,effective_radius_telescope,extinction_coef,output_beam_radius,stdv_fiber_lens_d,stdv_fiber_lens_offset,stdv_focal_length,unc_func):
+    def __init__(self,name,extinction_coef,unc_func):
                  self.ProbeVolumeID              = name         
-                 self.focal_length               = focal_length
-                 self.fiber_lens_d               = fiber_lens_d
-                 self.fiber_lens_offset          = fiber_lens_offset
-                 self.effective_radius_telescope = effective_radius_telescope
                  self.extinction_coef            = extinction_coef
-                 self.output_beam_radius         = output_beam_radius
-                 self.stdv_fiber_lens_d          = stdv_fiber_lens_d
-                 self.stdv_fiber_lens_offset     = stdv_fiber_lens_offset
-                 self.stdv_focal_length          = stdv_focal_length
                  self.Uncertainty                = unc_func
                  print('Class "Probe volume" created')
 #%% LIDAR MODULES
