+++ Utils/Qlunc_Plotting.py
@@ -11,7 +11,7 @@ University of Stuttgart(c)
 from Utils.Qlunc_ImportModules import *
 
 #%% Plotting:
-def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,flag_plot_photodetector_noise,flag_probe_volume_param):
+def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,flag_plot_photodetector_noise,flag_probe_volume_param,flag_plot_optical_amplifier_noise):
     """
     Plotting. Location: .Utils/Qlunc_plotting.py
     
@@ -92,9 +92,9 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
         a         = np.arange(2e-3,4e-3,.02e-3) # distance fiber-end--telescope lens
         a0        = Qlunc_yaml_inputs['Components']['Telescope']['Fiber-lens offset'] # the offset (a constant number), to avoid the fiber-end locates at the focal point, otherwise the lights will be parallel to each other
         A         = 20e-3 # beam radius at the output lens
-        ext_coef  = 0.085
+        ext_coef  = 1
         effective_radius_telescope  = 16.6e-3
-
+        s = 36.55 # distance from telescope to the target
         # The focus distance varies with the distance between the fiber-end and the telescope lens. So that, also the probe length varies with such distance.
         #Calculating focus distance depending on the distance between the fiber-end and the telescope lens:
         focus_distance = 1/((1/f_length)-(1/(a+a0))) # Focus distance
@@ -108,31 +108,39 @@ def plotting(Lidar,Qlunc_yaml_inputs,Data,flag_plot_measuring_points_pattern,fla
         vol_zr       = np.pi*(A**2)*(2*zr) # based on the definition of Rayleigh length in Liqin Jin notes (Focus calibration formula)
         
         # Lorentzian weighting function:
-        phi = (ext_coef/np.pi)*(1/((1**2)+(36.55-focus_distance)**2))
+        phi = (ext_coef/np.pi)*(zr/((zr**2)+(s-focus_distance)**2))
         
         # Plotting
         fig=plt.figure()
-        ax2=fig.add_subplot(2,2,1)
-        ax2.plot(dist,phi)
-        ax2.set_yscale('log')
+        axs2=fig.add_subplot(2,2,1)
+        axs2.plot(dist,phi)
+        axs2.set_yscale('log')
+        axs2.title.set_text('Weighting function')
+        axs2.set_xlabel('focus distance [m]',fontsize=plot_param['axes_label_fontsize'])
+        axs2.set_ylabel('$\phi$ [-]',fontsize=plot_param['axes_label_fontsize'])
+
+        axs3=fig.add_subplot(2,2,2)
+        axs3.plot(focus_distance,zr)
+        # axs3.set_xlabel('focus distance [m]',fontsize=plot_param['axes_label_fontsize'])
+        axs3.set_ylabel('{} [m]'.format('$\mathregular{z_{R}}$'),fontsize=plot_param['axes_label_fontsize'])
         
-            
-        # fig2=plt.figure()
-        ax3=fig.add_subplot(2,2,2)
-        ax3.plot(focus_distance,zr)
+        axs4=fig.add_subplot(2,2,3)
+        axs4.plot(a,zr)
+        axs4.set_xlabel('(a+a0) [m]',fontsize=plot_param['axes_label_fontsize'])
+        axs4.set_ylabel('{} [m]'.format('$\mathregular{z_{R}}$'),fontsize=plot_param['axes_label_fontsize'])
         
-        ax4=fig.add_subplot(2,2,3)
-        ax4.plot(a,zr)
         
[…]
