### 2021-08-07 b6d1107
Optical amplifier instance
