### 2021-08-07 fa552ad
Update rayleigh_distance.txt
