+++ metadata/rayleigh_distance.txt
@@ -1,1 +1,1 @@
-1.1392170823766439
+14.238114853496493
