+++ Main/Qlunc_inputs.yml
@@ -54,7 +54,7 @@ Components:
     Aperture: 2.e-2 #[m]
     Stdv Aperture: 2.e-6 #[m]   
     Focal length: 200.e-3
-    Fiber-lens distance: 3.5e-3
+    Fiber-lens distance: 2.35e-3
     Fiber-lens offset: 198.1e-3
     Effective radius telescope: 16.6e-3
     Output beam radius: 20.e-3
@@ -77,6 +77,7 @@ Components:
     Optical amplifier BW: 12.5e+12 # [Hz] Bandwidth optical amplifier (Hz)
     Optical amplifier noise figure: metadata/NoiseFigure.csv  #6# [dB] Relative Directory to where Noise figure data is. Either a well-known value or the directory where the file is.   
     Optical amplifier gain: 20 # [dB] Optical amplifier gain
+    Power interval: [0,1000,.01]       # [mW]
     Uncertainty function: uphc.UQ_Optical_amplifier # Python module! (You develop a function and save it as a python module)
 
   
@@ -163,7 +164,8 @@ Probe Volume:
 # Plotting:------------------------------------------------------------
 Flags:
   Pointing accuracy uncertainty: False
-  Scanning Pattern: True
+  Scanning Pattern: False
   Photodetector noise: False
   Save Scanning Coordinates: False
-  Probe Volume parameters: True
+  Probe Volume parameters: False
+  Optical_amplifier_noise: True
