### 2021-08-09 #3211 @PacoCosta
At a glance, I can see some people with knowledge of lidar and wind energy, which is what the topic is about. Also, python skills. These two seems suitable choices to me :
- Scivision (python, lidar, remote sensing, data science)
- antviro (python, wind power)
