+++ Main/Qlunc_inputs.yml
@@ -45,7 +45,7 @@ Components:
     
   Optical Circulator:
     Name: OC_Yaml
-    Insertion loss: 0.8 # [dB] # This is a term losses. Only have sense if the uncertainties are correlated...???
+    Insertion loss: 0.8 # [dB] # This is a term losses. Only Has sense if the uncertainties are correlated...???
     SNR: 60 #[dB]
     Uncertainty function: uopc.UQ_OpticalCirculator
   
@@ -54,11 +54,11 @@ Components:
     Aperture: 2.e-2 #[m]
     Stdv Aperture: 2.e-6 #[m]   
     Focal length: 200.e-3
-    Fiber-lens distance: 2.35e-3
+    Fiber-lens distance: 2.1e-3
     Fiber-lens offset: 198.1e-3
     Effective radius telescope: 16.6e-3
     Output beam radius: 20.e-3
-    stdv Focal length: 1.e-4
+    stdv Focal length: 1.e-6
     stdv Fiber-lens distance: 1.e-6
     stdv Fiber-lens offset: 1.e-6
     stdv Effective radius telescope: 1.e-6
@@ -121,23 +121,23 @@ Modules:
   Optics Module:
     Name: Optics_module_Yaml
     Scanner: Scanner
-    Optical circulator: Optical_circulator # Have to be the same name as the instance
+    Optical circulator: Optical_circulator # Has to be the same name as the instance
     Telescope: Telescope 
     Uncertainty function: uopc.sum_unc_optics # Python module! (You develop a function and save it as a python module)
   Photonics Module:
     Name: Photonics_module_Yaml
-    Photodetector: Photodetector           # Have to be the same name as the instance
-    Optical amplifier: Optical_Amplifier   # Have to be the same name as the instance
+    Photodetector: Photodetector           # Has to be the same name as the instance
+    Optical amplifier: Optical_Amplifier   # Has to be the same name as the instance
     Laser: Laser
     Uncertainty function: uphc.sum_unc_photonics # Python module! (You develop a function and save it as a python module)
 ---  
 # Lidar:
 Lidar:
   Name: Caixa1
-  Photonics module: Photonics_Module       # Have to be the same name as the instance
-  Optics module: None #Optics_Module             # Have to be the same name as the instance
-  Power module: None                       # Have to be the same name as the instance
-  Lidar inputs: Lidar_inputs               # Have to be the same name as the instance
+  Photonics module: Photonics_Module       # Has to be the same name as the instance
+  Optics module: None #Optics_Module             # Has to be the same name as the instance
+  Power module: None                       # Has to be the same name as the instance
+  Lidar inputs: Lidar_inputs               # Has to be the same name as the instance
   Uncertainty function: ulc.sum_unc_lidar  # Python module! (You develop a function and save it as a python module)
 
 ---
