### 2021-08-11 #3211 @danielskatz
Thanks @antviro - Given the delays up to this point, I'm in favor of a little more delay followed by a good review by a willing person than seeking someone else, I'm going assign you for now, and if someone else appears who can do this more quickly, I'll switch to that person.

I'll go ahead and invite you now, and if you can accept the invitation at https://github.com/openjournals/joss-reviews/invitations soon, it would be great - if not, it will expire and you will be unable to check off items when you are ready to start.
