+++ metadata/rayleigh_distance.txt
@@ -1,1 +1,1 @@
-14.238114853496493
+11.538627240062098

[…]
