### 2021-09-08 ff59175
SNR calculations
