### 2021-09-08 8eaf191
Introducing Active surface
Creating a new instance: Active surface in the photodetector
