+++ Main/Qlunc_Classes.py
@@ -103,13 +103,14 @@ class cts():
   
 # Component Classes:
 class photodetector():
-    def __init__(self,name,Photo_BandWidth,Load_Resistor,Photo_efficiency,Dark_Current,Photo_SignalP,Power_interval,Gain_TIA,V_Noise_TIA,unc_func):
+    def __init__(self,name,Photo_BandWidth,Load_Resistor,Photo_efficiency,Dark_Current,Photo_SignalP,Active_Surf,Power_interval,Gain_TIA,V_Noise_TIA,unc_func):
                  self.PhotodetectorID  = name 
                  self.BandWidth        = Photo_BandWidth 
                  self.Load_Resistor    = Load_Resistor 
                  self.Efficiency       = Photo_efficiency
                  self.DarkCurrent      = Dark_Current
                  self.SignalPower      = Photo_SignalP
+                 self.Active_Surf      = Active_Surf
                  self.Power_interval   = Power_interval
                  self.Gain_TIA         = Gain_TIA
                  self.V_Noise_TIA      = V_Noise_TIA
