+++ Main/Qlunc_inputs.yml
@@ -84,19 +84,20 @@ Components:
   
   Photodetector:
     Name: Photodetector_YAML
-    Photodetector BandWidth: 1.e+9    # [Hz]            
-    Load resistor: 50                  # [Ohm]
-    Photodetector efficiency: .85       # Photodetector efficiency [-]
+    Photodetector BandWidth: 100.e+6    # [Hz]            
+    Load resistor: 1.e+8                 # [Ohm]
+    Photodetector efficiency: .8       # Photodetector efficiency [-]
     Dark current: 5.e-9                 # In [A]. Dark current in the photodetector.
-    Photodetector signalP: 1.e-3        # [W]
-    Power interval: [1.e-6,1,1.e-4]           # [W]
+    Photodetector signalP: 1.e-2        # [W/m^2]
+    Power interval: [1.e-3,1.e-1,1.e-6]           # [W]
     Gain TIA: 5.e+3                     # [dB] Value or 'None' if no TIA in the photodetector
+    Active Surface: 1.e-6               #[m^2]
     V Noise TIA:  None # 160.e-6 #             # [V] Value or 'None' if no TIA in the photodetector
     Uncertainty function: uphc.UQ_Photodetector # Python module! (You develop a function and save it as a python module)
   
   Laser:
     Name: Laser_YAML
-    Wavelength: 1553.e-9    # [m]
+    Wavelength: 1000.e-9    # [m]
     stdv Wavelength: 1.e-9 #[m]
     Confidence interval: 3 # 1 => confidence interval= 68.27%, 2 => confidence interval= 90%, 3=> confidence interval= 95% 4=> confidence interval = 95.45%, 5 => confidence interval= 99%, 6=> confidence interval= 99.73%
     Output power: 12.5e-3 #[w]
@@ -112,7 +113,7 @@ Components:
   Lidar general inputs:
     Name: Lidar_gral_inputs_Yaml
     Type: CW # CW/Pulsed for continuous wave or pulsed lidars. The difference is that AOM (photonics module) is included in calculations when Pulsed
-    Wavelength: 1553.e-9    # [m]
+    Wavelength: 1000.e-9    # [m]
     Yaw error: 0       # [deg]
     Pitch error: 0         # [deg]
     Roll error: 0          # [deg]
