### 2021-09-15 #3211 @antviro
Hi @danielskatz,

Up to now, I have reviewed paper and the descriptions in the repository. They seem clear and quite interesting, I think they are ready to be published. Perhaps I miss some written documentation of the functions (besides the figure). Is there any? Despite there are two Jupyter notebooks describing its uses, It could be interesting.

Before finishing, I want to check functionality. I plan to do it within this week. I have no anaconda installation (I usually use pip instead, but perhaps I will install on other computer to check "official installation"). @PacoCosta: Are there instructions to use it without any package provider (besides installing the prerequisites)?

Best regards,
Antonio
