### 2021-09-16 #3211 @PacoCosta
Hi @antviro,
There aren't such instructions. If you have already installed python3.7 on your computer you can use pip to install the requirements, as you said. The good thing about using Anaconda is that you have a version of JupyterLab/Jupyter Notebook available there. Otherwise, if you want to go through the Tutorials, you have to install Jupyter separately.
You don't need to install anything else, but the requirements and Jupyter to run the code. Please remember to download the first release of the code, as stated in the _readme_ file on the repository.
Thanks for your time and effort.
