### 2021-09-23 #3211 @PacoCosta
Hi @antviro,
Yes, the problem is that new features have been added since the first release, where netcdf data conversion process was not needed at that point. When you download the first release "Qlunc-0.91" (and only this version), the code doesn't install this package since it is not in the requirements (_environment.yml_ file in the repository), and the code doesn't need it at all. I've just downloaded the first release and the installation process properly worked out for me following the instructions. If you use this version, you don't need to install netcdf. Anyway, I updated, as you suggested, the _environment.yml_ file in the latest code version with the netcdf version that the code installs. Please, let me know if, by downloading the first release "Qlunc-0.91",  and following instructions you have problems again; perhaps we can arrange a short meeting to go through the process together.
