### 2021-10-02 #3211 @danielskatz
@antviro - we can't publish until your checklist is complete - so how should we handle the item
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Can you state what specifically you think is needed so that the authors can work on it?
