+++ Docs/JOSS_202102/paper.bib
@@ -24,7 +24,7 @@
   	Author = {JCGM},
    ISBN = {9267101889},
   	Institution = {International Organization for Standardization},  	
-  	Title = {{Evaluation of measurement data â€” Guide to the expression of uncertainty in measurement}},
+  	Title = {{Evaluation of measurement data: Guide to the expression of uncertainty in measurement}},
   	Year = {2008}
 }
 
@@ -64,7 +64,7 @@
 
 @misc{mocalum,
   author       = {Vasiljevic, Nikola},
-  title        = {{mocalum: A python package for Monte-Carlo lidar 
+  title        = {{mocalum: A Python package for Monte-Carlo lidar 
                    uncertainty modeling}},
   month        = {may},
   year         = {2020},
