+++ Main/Qlunc_Instantiate.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-"""
+""".
+
 Created on Mon Oct 19 11:08:32 2020
 @author: fcosta
 
