+++ metadata/rayleigh_distance.txt
@@ -1,1 +1,1 @@
-11.538627240062098
+11.516337554472795
